% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-macros.R
\name{AddSeuratCommand}
\alias{AddSeuratCommand}
\title{Add a command to a Seurat object \verb{@commands} slot}
\usage{
AddSeuratCommand(
  object,
  name,
  call.string = paste0(name, "(object, ...)"),
  params = list(),
  assay.used = NULL,
  time.stamp = Sys.time()
)
}
\arguments{
\item{object}{Seurat object}

\item{name}{Name of the command}

\item{call.string}{Call string of the command}

\item{params}{Parameters of the command}

\item{assay.used}{Assay used in the command}

\item{time.stamp}{Time stamp of the command}
}
\value{
The Seurat object with the command added
}
\description{
Add a command to a Seurat object \verb{@commands} slot
}
\examples{
object <- SeuratObject::pbmc_small
names(object@commands)

object <- AddSeuratCommand(
    object, "RunSeuratDEAnalysis",
    "RunSeuratDEAnalysis(object, group_by = 'groups', ident_1 = 'g1', ident_2 = 'g2')",
    list(group_by = "groups", ident_1 = "g1", ident_2 = "g2")
)
object@commands$RunSeuratDEAnalysis
}
