% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biopipen.utils-package.R
\docType{package}
\name{biopipen.utils-package}
\alias{biopipen.utils}
\alias{biopipen.utils-package}
\title{biopipen.utils: Common Utility Functions for Biopipen}
\description{
Common utility used for biopipen processes to compose bioinformatics
pipelines.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pwwang/biopipen.utils.R https://pwwang.github.io/biopipen.utils.R/}
  \item Report bugs at \url{https://github.com/pwwang/biopipen.utils.R/issues}
}

}
\author{
\strong{Maintainer}: Panwen Wang \email{pwwang@pwwang.com} (\href{https://orcid.org/0000-0002-4614-8970}{ORCID})

}
\keyword{internal}
