% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{bQuote}
\alias{bQuote}
\title{backtick quoting}
\usage{
bQuote(x, force = FALSE)
}
\arguments{
\item{x}{a character vector}

\item{force}{whether to force backtick quoting}
}
\value{
backtick quoted character vector
}
\description{
backtick quoting
}
\examples{
bQuote(c("a", "b", "c"))
# [1] "a" "b" "c"
bQuote(c("a 1", "b 2", "c 3"))
# [1] "`a 1`" "`b 2`" "`c 3`"
}
