% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-interop.R
\name{ConvertAnnDataToSeurat}
\alias{ConvertAnnDataToSeurat}
\title{Convert an AnnData file (h5ad) to a Seurat object or an RDS/qs2 file}
\usage{
ConvertAnnDataToSeurat(
  infile,
  outfile = NULL,
  assay = NULL,
  ident = NULL,
  log = NULL
)
}
\arguments{
\item{infile}{Input file}

\item{outfile}{Output file. If "NULL" (default) is given, a temporary file will be created
and saved and the Seurat object will be read from it.}

\item{assay}{Assay name to save in the Seurat object}

\item{ident}{The name of the identity in metadata after conversion}

\item{log}{Logger}
}
\value{
The Seurat object if \code{outfile} is "\if{html}{\out{<object>}}", otherwise NULL.
}
\description{
Convert an AnnData file (h5ad) to a Seurat object or an RDS/qs2 file
}
