% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat-disk-patch.R
\name{.AssembleAssay}
\alias{.AssembleAssay}
\title{Patch for SeuratDisk's AssembleAssay function}
\usage{
.AssembleAssay(assay, file, slots = NULL, verbose = TRUE)
}
\description{
Patch for SeuratDisk's AssembleAssay function
}
\keyword{internal}
