% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{.dir.copy}
\alias{.dir.copy}
\title{Copy a directory to another location}
\usage{
.dir.copy(from, to, overwrite = FALSE, copy.date = TRUE)
}
\arguments{
\item{from}{Source directory path}

\item{to}{Target directory path}

\item{overwrite}{Whether to overwrite existing files in the target directory}

\item{copy.date}{Whether to copy the file modification date}
}
\value{
a logical vector indicating which operation succeeded for each of the files in \code{from} attempted
}
\description{
Unlike \code{file.copy}, this function copies a directory and its contents
to another location, instead of copying into another location.
It creates the target directory if it does not exist.
}
\keyword{internal}
