% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-macros.R
\name{.expand_findclusters_resolution}
\alias{.expand_findclusters_resolution}
\title{Expand the resolution for \code{FindClusters}}
\usage{
.expand_findclusters_resolution(resolution)
}
\arguments{
\item{resolution}{A numeric vector or a character vector}
}
\value{
A numeric vector with the expanded resolution
}
\description{
So that we can have 0.1:0.4:0.1 to be expanded to
0.1, 0.2, 0.3, 0.4
}
\keyword{internal}
