% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-macros.R
\name{.expand_number}
\alias{.expand_number}
\title{AddSeuratCommand
Expand a single numer into 1:number}
\usage{
.expand_number(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric vector
}
\description{
This is useful to expand args$dims = 10 to args$dims = 1:10
}
\keyword{internal}
