% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5ad_to_h5seurat.R
\name{.H5ADToH5Seurat}
\alias{.H5ADToH5Seurat}
\title{patch to fix https://github.com/mojaveazure/seurat-disk/issues/31}
\usage{
.H5ADToH5Seurat(source, dest, assay = "RNA", overwrite = FALSE, verbose = TRUE)
}
\description{
patch to fix https://github.com/mojaveazure/seurat-disk/issues/31
}
\keyword{internal}
