% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat-disk-patch.R
\name{.H5SeuratToH5AD}
\alias{.H5SeuratToH5AD}
\title{patch to fix feature names not attached to the raw data}
\usage{
.H5SeuratToH5AD(
  source,
  dest,
  assay = DefaultAssay(object = source),
  overwrite = FALSE,
  verbose = TRUE
)
}
\description{
patch to fix feature names not attached to the raw data
}
\keyword{internal}
