% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-macros.R
\name{.recode_clusters}
\alias{.recode_clusters}
\title{Recode the clusters, from 0, 1, 2, ... to x1, x2, x3, ...}
\usage{
.recode_clusters(clusters, prefix = "c")
}
\arguments{
\item{clusters}{A numeric vector of clusters}
}
\value{
A character vector of recoded clusters
}
\description{
Recode the clusters, from 0, 1, 2, ... to x1, x2, x3, ...
}
\keyword{internal}
