% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-macros.R
\name{.rename_seurat_features}
\alias{.rename_seurat_features}
\title{Rename features in a Seurat object}
\usage{
.rename_seurat_features(obj, features_map)
}
\arguments{
\item{obj}{A Seurat object}

\item{features_map}{A named character vector, where the names are the original feature names and the values are the new names}
}
\value{
A Seurat object with renamed features
}
\description{
Rename features in a Seurat object
}
\keyword{internal}
