% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{.run_edger}
\alias{.run_edger}
\title{Run edgeR differential expression analysis}
\usage{
.run_edger(exprs, meta, group_by, ident_1, ident_2, paired_by, log, ncores)
}
\arguments{
\item{exprs}{Expression matrix with genes as rows and samples as columns.}

\item{meta}{Metadata data frame with sample information.}

\item{group_by}{Column name in \code{meta} to group samples for differential expression analysis}

\item{ident_1}{First identity to compare against}

\item{ident_2}{Second identity to compare against}

\item{paired_by}{Column name in \code{meta} for paired samples.}

\item{log}{Logger}

\item{ncores}{Number of cores to use for parallel processing, not used in this function.}
}
\value{
A data frame with edgeR differential expression results.
}
\description{
Run edgeR differential expression analysis
}
\keyword{internal}
