% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{.sig_str}
\alias{.sig_str}
\title{Generate a string representation of an R object for caching signature}
\usage{
.sig_str(object)
}
\arguments{
\item{object}{The R object to generate the string representation for}
}
\value{
A character vector representing the structure of the object
}
\description{
This function generates a string representation of an R object using \code{utils::str()}.
The output is designed to be consistent and comprehensive for caching purposes.
It captures the structure of the object, including its attributes, and is used to create a unique signature for caching.
}
\keyword{internal}
