% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-helper.R
\name{.size_compare}
\alias{.size_compare}
\title{Internal helper function to compare the size of two groups.}
\usage{
.size_compare(
  df,
  group_by,
  idents,
  id,
  fun,
  compare = ".n",
  split_by = NULL,
  order = "desc(sum)",
  subset = NULL,
  return_type = c("uids", "ids", "subdf", "df", "interdf"),
  include_zeros = FALSE
)
}
\arguments{
\item{df}{The data frame. Use \code{.} if the function is called in a dplyr pipe.}

\item{group_by}{The column name in the data frame to group the entities.
It could be a quoted string or a bare variable, and defines the groups of entities
for comparison.}

\item{idents}{The values in \code{compare} to compare. It could be either an
an integer or a vector. If it is an integer, the number of values in
\code{compare} must be the same as the integer for the \code{id} to be regarded
as paired. If it is a vector, the values in \code{compare} must be the same
as the values in \code{idents} for the \code{id} to be regarded as paired.}

\item{id}{The column name in \code{df} for the groups.}

\item{fun}{The way to compare between groups. Either \code{"more_than"},
\code{"less_than"}, \code{"emerged"} or \code{"vanished"}.}

\item{compare}{The column name in \code{df} to compare the values for each group.
It could be either a numeric column or \code{.n} to compare the number of
entities in each group. If a column is passed, the values in the column
must be numeric and the same in each group. This won't be checked.}

\item{split_by}{A column name (without quotes) in metadata to split the cells.}

\item{order}{An expression to order the intermediate data frame before returning
the final result. Default is \code{NULL}. It does not work for \code{subdf} and \code{df}.}

\item{subset}{An expression to subset the entities, will be passed to
\code{dplyr::filter()}. Default is \code{TRUE} (no filtering).}

\item{return_type}{The type of the returned value. Default is \code{uids}.
It could be one of
\itemize{
\item \code{uids}: return the unique ids of the selected entities
\item \code{ids}: return the ids of all entities in the same order as in \code{df}, where the
non-selected ids will be \code{NA}
\item \code{subdf}: return a subset of \code{df} with the selected entities
\item \code{df}: return the original \code{df} with a new logical column \code{.out} to mark
the selected entities
\item \code{interdf}: return the intermediate data frame with the id column, \verb{<compare>},
\code{predicate} and the split_by column if provided.
}}

\item{include_zeros}{Whether to include the zero entities in the other group for
\code{more_than} and \code{less_than} comparisons. Default is \code{FALSE}.
By default, the zero entities will be excluded, meaning that the entities must
exist in both groups to be selected.}
}
\value{
A vector of ids or a data frame with the selected entities.
}
\description{
Internal helper function to compare the size of two groups.
}
\keyword{internal}
