% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casing.R
\name{expand_cases}
\alias{expand_cases}
\title{Expand the cases with default values}
\usage{
expand_cases(cases, defaults, post = NULL, default_case = "DEFAULT")
}
\arguments{
\item{cases}{A list of cases}

\item{defaults}{A list of default values}

\item{post}{A function for post-handling each case, returning the a list with the
name and the case. One can also return multiple cases based on the demand.
All cases returned will be finally merged.}

\item{default_case}{The name of the default case}
}
\value{
A list of expanded cases
}
\description{
Expand the cases with default values
}
