% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{format_args}
\alias{format_args}
\title{Format the arguments for debugging}
\usage{
format_args(args)
}
\arguments{
\item{args}{The arguments to format}
}
\value{
The formatted arguments
}
\description{
Format the arguments for debugging
}
\keyword{internal}
