% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{gene_name_conversion}
\alias{gene_name_conversion}
\title{Convert gene names between different formats}
\usage{
gene_name_conversion(
  genes,
  infmt,
  outfmt,
  dup = "first",
  species = "human",
  notfound = "na",
  suppress_messages = TRUE
)
}
\arguments{
\item{genes}{A character/integer vector of gene names/ids}

\item{infmt}{A character vector of input gene name formats
See the available scopes at
https://docs.mygene.info/en/latest/doc/data.html#available-fields
You can use ensg as a shortcut for ensembl.gene}

\item{outfmt}{A character vector of output gene name formats}

\item{dup}{How to deal with duplicate gene names found.
\itemize{
\item "first": keep the first one (default), sorted by score descendingly
\item "last": keep the last one, sorted by score descendingly
\item "all": keep all of them, each will be a separate row
\item "\verb{<X>}": combine them into a single string, separated by \code{X}
}}

\item{species}{A character vector of species names}

\item{notfound}{How to deal with gene names that are not found
\itemize{
\item "error": stop with an error message
\item "use-query": use the query gene name as the converted gene name
\item "skip": skip the gene names that are not found
\item "ignore": Same as "skip"
\item "na"/"NA": use NA as the converted gene name (default)
}}

\item{suppress_messages}{Whether to suppress the warning messages}
}
\value{
A tibble with the query gene names and the converted gene names
When a gene name is not found, the converted name will be NA
When duplicate gene names are found, the one with the highest score will be kept
}
\description{
Convert gene names between different formats
}
