% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_biopipen_dir}
\alias{get_biopipen_dir}
\title{Get the path where biopipen is installed}
\usage{
get_biopipen_dir(python = "python")
}
\arguments{
\item{python}{The path to the Python executable at runtime.}
}
\value{
The path where biopipen is installed
}
\description{
This is useful when the script is running in a different environment/platform.
At runtime, we need to know where the biopipen package is installed so that
we can source or use some files (or scripts) shipped with biopipen.
}
