% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{get_cached}
\alias{get_cached}
\title{Get signatures and cached data}
\usage{
get_cached(x, kind, cache_dir)
}
\arguments{
\item{x}{An object to infer signature from}

\item{kind}{A string indicating the kind of the object
Used as part of the filename of the cached file}

\item{cache_dir}{A string indicating the directory to store cached files}
}
\value{
A list containing the signature, digested signature and cached data
}
\description{
Get signatures and cached data
}
\details{
This function is used to get the signature of an object and the cached data
if it exists. The signature is used to identify the object and the cached data
is used to avoid recomputing the object.
}
