% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{get_logger}
\alias{get_logger}
\title{Setup and return the logger}
\usage{
get_logger(
  format =
    "{sprintf(\\"\%-7s\\", level)} [{format(time, \\"\%Y-\%m-\%d \%H:\%M:\%S\\")}] {msg}",
  appender = "stdout",
  level = "INFO"
)
}
\arguments{
\item{format}{The format of the log message.
This format will be passed to \link[logger:layout_glue_generator]{logger::layout_glue_generator}. The default format will generate a log message like:
\verb{INFO    [2021-01-01 00:00:00] Hello world}}

\item{appender}{The appender to use. Default is "stdout"
Either a string or a logger appender object. If string is provided, it will be converted to the corresponding logger appender object.
For example, "stdout" will be converted to \link[logger:appender_stdout]{logger::appender_stdout}}

\item{level}{The log level. Default is "INFO"
Either a string or a logger level object. If string is provided, it will be converted to the corresponding logger level object.
For example, "INFO" will be converted to \link[logger:log_levels]{logger::INFO}}
}
\value{
A list of logging functions
}
\description{
Setup and return the logger
}
