% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_worh}
\alias{get_worh}
\title{Get the height or width of the plot by given default size and user-specified size}
\usage{
get_worh(default_size, user_size)
}
\arguments{
\item{default_size}{The default size of the plot}

\item{user_size}{The user-specified size of the plot, can be NULL, a single number,
or "+", "-", "*", "/" followed by a number to indicate relative size to the default size.}
}
\value{
The calculated size of the plot
}
\description{
Get the height or width of the plot by given default size and user-specified size
}
\examples{
\dontrun{
get_worh(5, NULL)  # returns 5
get_worh(5, 10)    # returns 10
get_worh(5, "+2")  # returns 7
get_worh(5, "-1")  # returns 4
get_worh(5, "*2")  # returns 10
get_worh(5, "/2")  # returns 2.5
}
}
\keyword{internal}
