% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{gettempdir}
\alias{gettempdir}
\title{Get the temporary directory, without suffix.
It works like \code{tempfile.gettempdir()} in Python.}
\usage{
gettempdir()
}
\value{
The temporary directory
}
\description{
Get the temporary directory, without suffix.
It works like \code{tempfile.gettempdir()} in Python.
}
