% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scps.R
\name{GS_Experssion_Calculation}
\alias{GS_Experssion_Calculation}
\title{Calculate the mean expression level of gene sets}
\usage{
GS_Experssion_Calculation(Seurat_data, GeneSet)
}
\arguments{
\item{Seurat_data}{A Seurat object (Seurat V5) with a normalized \code{SCT}
or \code{RNA} assay}

\item{GeneSet}{A \code{GeneSet} or a \code{GeneSetCollection}}
}
\value{
A matrix of mean expression levels, gene sets x cells
}
\description{
Ported from the scPS script (https://github.com/Thakar-Lab/scPS).
}
\keyword{internal}
