% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5ad_to_h5seurat.R
\name{h5group_to_list}
\alias{h5group_to_list}
\title{Convert H5D or H5Group to a list}
\usage{
h5group_to_list(h5group)
}
\arguments{
\item{h5group}{An H5D or H5Group object}
}
\value{
A list representation of the H5D or H5Group
}
\description{
When the structure of the HDF5 file is miscellaneous, simply
using \code{as.list()} will not work.
}
\keyword{internal}
