% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-markers.R
\name{is_marker_canonical}
\alias{is_marker_canonical}
\title{Check whether a marker table has the canonical columns}
\usage{
is_marker_canonical(df)
}
\arguments{
\item{df}{A marker table.}
}
\value{
\code{TRUE} if \code{df} is a data.frame with \code{cell_type} and \code{gene} columns,
\code{FALSE} otherwise.
}
\description{
A canonical marker table has both the \code{cell_type} and \code{gene}
columns.
}
