% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{list_rename}
\alias{list_rename}
\title{Rename to names of a list}
\usage{
list_rename(x, fn)
}
\arguments{
\item{x}{A list}

\item{fn}{A function to rename the names.
If the function returns NULL or FALSE, the name will be removed.
If the function returns TRUE, the name will be kept.
If the function returns a string, the name will be changed to the string.
The function can take 1 or 2 arguments, the first is the name, the second is the value.}
}
\value{
The list with renamed names
}
\description{
Rename to names of a list
}
