% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5ad_to_h5seurat.R
\name{list_to_h5group}
\alias{list_to_h5group}
\title{Then write the list to a group}
\usage{
list_to_h5group(h5fg, name, lst)
}
\arguments{
\item{h5fg}{An H5File or H5Group object where the group will be created}

\item{name}{The name of the group to create}

\item{lst}{The list to write to the group}
}
\value{
The created H5Group object
}
\description{
This function writes a list to an H5Group, creating datasets for each element.
If the list contains nested lists, it will create subgroups.
}
\keyword{internal}
