% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{list_update}
\alias{list_update}
\title{Update the first list based on other lists}
\usage{
list_update(x, ..., depth = -1L)
}
\arguments{
\item{x}{The first list}

\item{...}{The other lists}

\item{depth}{The depth to update, -1 means update all}
}
\value{
The updated list
}
\description{
Update the first list based on other lists
}
