% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{list_update2}
\alias{list_update2}
\title{Update a list with another list}
\usage{
list_update2(x, y, depth = -1L)
}
\arguments{
\item{x}{A list}

\item{y}{A list}

\item{depth}{The depth to update, -1 means update all}
}
\value{
The updated list
}
\description{
Update a list with another list
}
\keyword{internal}
