% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-load-qc.R
\name{LoadSeuratAndPerformQC}
\alias{LoadSeuratAndPerformQC}
\title{Load samples into a Seurat object}
\usage{
LoadSeuratAndPerformQC(
  meta,
  min_cells = 0,
  min_features = 0,
  features = NULL,
  samples = NULL,
  cell_qc = NULL,
  gene_qc = NULL,
  ccs_args = NULL,
  LoadLoomArgs = list(),
  contam_correction = NULL,
  decontXArgs = list(),
  scCDCArgs = list(Detection = list(), Quantification = list(), Correction = list()),
  keep_contam_assay = FALSE,
  tmpdir = NULL,
  log = NULL,
  cache = NULL
)
}
\arguments{
\item{meta}{Metadata of the samples
Required columns: Sample, RNAData.
The RNAData column should contain the path to the 10X or ParseBio data, either a directory or a file
If the path is a directory, the function will look for barcodes.tsv.gz, features.tsv.gz and matrix.mtx.gz.
The directory should be loaded by \link[Seurat:Read10X]{Seurat::Read10X}, \link[Seurat:ReadParseBio]{Seurat::ReadParseBio} or the HIVE data. Sometimes, there may be prefix in the file names,
e.g. "'prefix'.barcodes.tsv.gz", which is also supported.
If the path is a file ending with ".loom", it will be loaded by \code{\link[SeuratDisk:Connect]{SeuratDisk::Connect()}} and converted to a Seurat object.
Otherwise, if the path is a file, it should be a h5 file that can be loaded by \code{\link[Seurat:Read10X_h5]{Seurat::Read10X_h5()}}

This can also be a Seurat object to perform QC on. It requires the "Sample" column in the meta.data slot
specifying the sample for each cell.}

\item{min_cells}{Include features detected in at least this many cells.
This will be applied to all samples and passed to the \code{\link[Seurat:reexports]{Seurat::CreateSeuratObject()}} function.
QCs can be further performed on the object after loading.
You can also provide a list of values, where the names of the list are sample names
and the values are the minimum number of cells for each sample to load by
\code{\link[Seurat:reexports]{Seurat::CreateSeuratObject()}}.
You can have a default value in the list with the name "DEFAULT" for the samples
that are not listed.
This won't work if data is loaded from a loom file or \code{meta} is a Seurat object.}

\item{min_features}{Include cells where at least this many features are detected.
This will be applied to all samples and passed to the \code{\link[Seurat:reexports]{Seurat::CreateSeuratObject()}} function.
QCs can be further performed on the object after loading.
You can also provide a list of values, where the names of the list are sample names
and the values are the minimum number of features for each sample to load by
\code{\link[Seurat:reexports]{Seurat::CreateSeuratObject()}}.
You can have a default value in the list with the name "DEFAULT" for the samples
that are not listed.
This won't work if data is loaded from a loom file or \code{meta} is a Seurat object.}

\item{features}{A named character vector/list or a file path to rename features.
If a named vector/list is given, the names are the original feature names and the values are
the new names.
If a file path is given, it should be a TAB-delimited file with two columns (no header);
lines beginning with '#' are ignored. The first column contains the original feature names
and the second column the new names.}

\item{samples}{Samples to load. If NULL, all samples will be loaded}

\item{cell_qc}{Cell QC criteria, a string of expression to pass to \code{dplyr::filter} function
to filter the cells.
It can also be a list of expressions, where the names of the list are sample names.
You can have a default expression in the list with the name "DEFAULT" for the samples
that are not listed.}

\item{gene_qc}{Gene QC criteria
A list containing the following fields:
\itemize{
\item min_cells: Minimum number of cells a gene should be expressed in to be kept
\item excludes: A string or strings to exclude certain genes. Regular expressions are supported.
Multiple strings can also be separated by commas in a single string.
}}

\item{ccs_args}{Arguments to pass to \code{\link[=RunSeuratCellCycleScoring]{RunSeuratCellCycleScoring()}} to perform cell cycle scoring after the data
has been QC'ed.}

\item{LoadLoomArgs}{Arguments to pass to \code{\link[SeuratDisk:LoadLoom]{SeuratDisk::LoadLoom()}} when loading loom files.}

\item{contam_correction}{Whether to perform contaminant RNA correction. Supported methods: "decontx" (using decontX function from the celda package) and "sccdc" (using scCDC package).}

\item{decontXArgs}{Arguments to pass to decontX function from the celda package when \code{contam_correction} is "decontx". See \code{?celda::decontX} for details.}

\item{scCDCArgs}{Arguments to pass to scCDC package when \code{contam_correction} is "sccdc". It should be a list containing three sub-lists: Detection, Quantification and Correction, which are the arguments for \code{scCDC::ContaminationDetection}, \code{scCDC::ContaminationQuantification} and \code{scCDC::ContaminationCorrection} functions, respectively. See \code{?scCDC::ContaminationDetection}, \code{?scCDC::ContaminationQuantification} and \code{?scCDC::ContaminationCorrection} for details.}

\item{keep_contam_assay}{Whether to keep the \code{Contaminated} assay (the original counts before
contamination correction) in the object. If \code{FALSE} (default), the assay is dropped per-sample
right after the correction (before samples are merged) to save memory. Note the \code{Contaminated}
assay is not used by any downstream steps of \code{\link[=FinishSeuratQC]{FinishSeuratQC()}} unless \code{keep_contam_assay} is \code{TRUE}.}

\item{tmpdir}{Temporary directory to store intermediate files when there are prefix in the file names}

\item{log}{Logger}

\item{cache}{Directory to cache the results. Set to \code{FALSE} to disable caching}
}
\value{
A Seurat object.
In slot \code{misc}, there are two fields: \code{cell_qc_df} and \code{gene_qc}, a list of gene QC results
}
\description{
Cell QC will be performed, either per-sample or on the whole object
}
\examples{
\donttest{
datadir <- system.file("extdata", "scrna", package = "biopipen.utils")
meta <- data.frame(
    Sample = c("Sample1", "Sample2"),
    RNAData = c(
        file.path(datadir, "Sample1"),
        file.path(datadir, "Sample2")
    )
)

obj <- LoadSeuratAndPerformQC(
   meta, cache = FALSE, cell_qc = "nFeature_RNA > 1000",
   gene_qc = list(min_cells = 3)
)
print(table(obj$.QC))
print(obj@misc$gene_qc)
}
}
