% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-load-qc.R
\name{LoadSeuratSamples}
\alias{LoadSeuratSamples}
\title{Load samples into a list of Seurat objects}
\usage{
LoadSeuratSamples(
  meta,
  min_cells = 0,
  min_features = 0,
  features = NULL,
  samples = NULL,
  LoadLoomArgs = list(),
  tmpdir = NULL,
  log = NULL
)
}
\arguments{
\item{meta}{Metadata of the samples
Required columns: Sample, RNAData.
The RNAData column should contain the path to the 10X or ParseBio data, either a directory or a file
If the path is a directory, the function will look for barcodes.tsv.gz, features.tsv.gz and matrix.mtx.gz.
The directory should be loaded by \link[Seurat:Read10X]{Seurat::Read10X}, \link[Seurat:ReadParseBio]{Seurat::ReadParseBio} or the HIVE data. Sometimes, there may be prefix in the file names,
e.g. "'prefix'.barcodes.tsv.gz", which is also supported.
If the path is a file ending with ".loom", it will be loaded by \code{\link[SeuratDisk:Connect]{SeuratDisk::Connect()}} and converted to a Seurat object.
Otherwise, if the path is a file, it should be a h5 file that can be loaded by \code{\link[Seurat:Read10X_h5]{Seurat::Read10X_h5()}}

This can also be a Seurat object to split into samples. It requires the "Sample" column in the meta.data slot
specifying the sample for each cell.}

\item{min_cells}{Include features detected in at least this many cells.
This will be applied to all samples and passed to the \code{\link[Seurat:reexports]{Seurat::CreateSeuratObject()}} function.
QCs can be further performed on the object after loading.
You can also provide a list of values, where the names of the list are sample names
and the values are the minimum number of cells for each sample to load by
\code{\link[Seurat:reexports]{Seurat::CreateSeuratObject()}}.
You can have a default value in the list with the name "DEFAULT" for the samples
that are not listed.
This won't work if data is loaded from a loom file or \code{meta} is a Seurat object.}

\item{min_features}{Include cells where at least this many features are detected.
This will be applied to all samples and passed to the \code{\link[Seurat:reexports]{Seurat::CreateSeuratObject()}} function.
QCs can be further performed on the object after loading.
You can also provide a list of values, where the names of the list are sample names
and the values are the minimum number of features for each sample to load by
\code{\link[Seurat:reexports]{Seurat::CreateSeuratObject()}}.
You can have a default value in the list with the name "DEFAULT" for the samples
that are not listed.
This won't work if data is loaded from a loom file or \code{meta} is a Seurat object.}

\item{features}{A named character vector/list or a file path to rename features.
If a named vector/list is given, the names are the original feature names and the values are
the new names.
If a file path is given, it should be a TAB-delimited file with two columns (no header);
lines beginning with '#' are ignored. The first column contains the original feature names
and the second column the new names.}

\item{samples}{Samples to load. If NULL, all samples will be loaded}

\item{LoadLoomArgs}{Arguments to pass to \code{\link[SeuratDisk:LoadLoom]{SeuratDisk::LoadLoom()}} when loading loom files.}

\item{tmpdir}{Temporary directory to store intermediate files when there are prefix in the file names}

\item{log}{Logger}
}
\value{
A named list of Seurat objects, one per sample. Samples that have no data or no
cells are skipped with a warning.
}
\description{
This loads each sample without performing any QC, so that QC can be done per sample by
\code{\link[=PerformSeuratQC]{PerformSeuratQC()}} before the samples are merged by \code{\link[=LoadSeuratAndPerformQC]{LoadSeuratAndPerformQC()}}.
}
\examples{
\donttest{
datadir <- system.file("extdata", "scrna", package = "biopipen.utils")
meta <- data.frame(
    Sample = c("Sample1", "Sample2"),
    RNAData = c(
        file.path(datadir, "Sample1"),
        file.path(datadir, "Sample2")
    )
)

objs <- LoadSeuratSamples(meta)
names(objs)
}
}
