% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-markers.R
\name{markers_to_garnett_file}
\alias{markers_to_garnett_file}
\title{Convert a marker table to a garnett-native marker file}
\usage{
markers_to_garnett_file(df, file, tissue = NULL, cancer = NULL, species = NULL)
}
\arguments{
\item{df}{A canonical marker table.}

\item{file}{The path of the garnett marker file to write.}

\item{tissue}{The tissue to filter by. If NULL, no filtering is done.}

\item{cancer}{The cancer to filter by. If NULL, no filtering is done.}

\item{species}{The species to filter by. If NULL, no filtering is done.}
}
\value{
The path of the written garnett marker file.
}
\description{
Write \verb{> <cell type>} blocks with \verb{expressed:} /
\verb{not expressed:} rules. Negative-direction markers become
\verb{not expressed:} rules.
}
