% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{monkey_patch}
\alias{monkey_patch}
\title{Monkey patch a function from a namespace}
\usage{
monkey_patch(namespace, function_name, new_function)
}
\arguments{
\item{namespace}{The namespace where the function is located}

\item{function_name}{The name of the function to be patched}

\item{new_function}{The new function to be patched}
}
\description{
Monkey patch a function from a namespace
}
\keyword{internal}
