% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-helper.R
\name{paired}
\alias{paired}
\title{Get paired entities from a data frame based on the other column}
\usage{
paired(df = ".", id, compare, idents = 2, uniq = TRUE)
}
\arguments{
\item{df}{The data frame.}

\item{id}{The column name in \code{df} for the ids to be returned in the
final output}

\item{compare}{The column name in \code{df} to compare the values for each
id in \code{id}.}

\item{idents}{The values in \code{compare} to compare. It could be either an
an integer or a vector. If it is an integer, the number of values in
\code{compare} must be the same as the integer for the \code{id} to be regarded
as paired. If it is a vector, the values in \code{compare} must be the same
as the values in \code{idents} for the \code{id} to be regarded as paired.}

\item{uniq}{Whether to return unique ids or not. Default is \code{TRUE}.
If \code{FALSE}, you can mutate the meta data frame with the returned ids.
Non-paired ids will be \code{NA}.}
}
\value{
A vector of paired ids (in \code{id} column)
}
\description{
Get paired entities from a data frame based on the other column
}
\examples{
df <- data.frame(
    id = c("A", "A", "B", "B", "C", "C", "D", "D"),
    compare = c(1, 2, 1, 1, 1, 2, 1, 2)
)
paired(df, id, compare, 2)
paired(df, id, compare, c(1, 2))
paired(df, id, compare, c(1, 2), uniq = FALSE)
}
