% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-helper.R
\name{top}
\alias{top}
\title{Get top entities from a data frame based on the number of entities in each group}
\usage{
top(
  df = ".",
  id,
  n = 10,
  compare = ".n",
  subset = NULL,
  with_ties = FALSE,
  split_by = NULL,
  return_type = c("uids", "ids", "subdf", "df", "interdf")
)
}
\arguments{
\item{df}{The data frame. Use \code{.} if the function is called in a dplyr pipe.}

\item{id}{The column name in \code{df} for the groups.}

\item{n}{The number of top entities to return. if \code{n} < 1, it will be
regarded as the percentage of the total number of entities in each group
(after subsetting or each applied).
Specify 0 to return all entities.}

\item{compare}{The column name in \code{df} to compare the values for each group.
It could be either a numeric column or \code{.n} to compare the number of
entities in each group. If a column is passed, the values in the column
must be numeric and the same in each group. This won't be checked.}

\item{subset}{An expression to subset the entities, will be passed to
\code{dplyr::filter()}. Default is \code{TRUE} (no filtering).}

\item{with_ties}{Whether to return all entities with the same size as the
last entity in the top list. Default is \code{FALSE}.}

\item{split_by}{A column name (without quotes) in metadata to split the cells.}

\item{return_type}{The type of the returned value. Default is \code{uids}.
It could be one of
\itemize{
\item \code{uids}: return the unique ids of the selected entities
\item \code{ids}: return the ids of all entities in the same order as in \code{df}, where the
non-selected ids will be \code{NA}
\item \code{subdf}: return a subset of \code{df} with the selected entities
\item \code{df}: return the original \code{df} with a new logical column \code{.out} to mark
the selected entities
\item \code{interdf}: return the intermediate data frame with the id column, \verb{<compare>},
\code{predicate} and the split_by column if provided.
}}
}
\value{
Depending on the \code{return_type}, the function will return different values.
}
\description{
Get top entities from a data frame based on the number of entities in each group
}
\examples{
df <- data.frame(
    id = c("A", "B", "C", "D", "E", "F", "G", "H"),
    value = c(10, 20, 30, 40, 50, 60, 80, 80)
)
top(df, id, n = 1, compare = value, with_ties = TRUE, return_type = "uids")
top(df, "id", n = 2, compare = "value", return_type = "subdf")
top(df, "id", n = 2, compare = "value", return_type = "df")
top(df, "id", n = 2, compare = "value", return_type = "interdf")
top(df, id, n = 0.25, compare = value, return_type = "uids")
top(df, id, n = 0, compare = value, return_type = "uids")

df <- data.frame(id = c("A", "A", "B", "B", "B", "C", "C", "C", "D", "D", "D", "D"))
top(df, id, n = 2, compare = ".n", return_type = "uids", with_ties = TRUE)
dplyr::mutate(df, selected = top(id = id, n = 2, compare = ".n", return_type = "ids",
  with_ties = TRUE))
}
