% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctcr-marcos.R
\name{MutateScRep}
\alias{MutateScRep}
\alias{ScRepMutate}
\title{Mutate scRepertorie object}
\usage{
MutateScRep(screp, mutaters, log = NULL)

ScRepMutate(screp, mutaters, log = NULL)
}
\arguments{
\item{screp}{The scRepertorie object. It is either a Seurat object or a list of data.frames}

\item{mutaters}{A named list of mutater expressions, where the names are the new column names
and the values are the expressions to mutate the columns
The name with the suffix \verb{:ident} will be used as the new identity column (only for Seurat object)
The values can be either character strings of expressions to be parsed}

\item{log}{Logger object to log the messages. If NULL, the default logger will be used.}
}
\value{
The mutated scRepertorie object
}
\description{
Mutate scRepertorie object

Alias of MutateScRep
}
\examples{
\donttest{
data(contig_list, package = "scRepertoire")
screp <- scRepertoire::combineTCR(contig_list)
head(MutateScRep(screp, list(CTaa_len = "nchar(CTaa)"))[[1]])
}
}
\seealso{
\code{\link[=MutateSeuratMeta]{MutateSeuratMeta()}}
}
