% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-macros.R
\name{MutateSeuratMeta}
\alias{MutateSeuratMeta}
\title{Mutater the Seurat metadata}
\usage{
MutateSeuratMeta(object, mutaters, log = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{mutaters}{A named list of mutater expressions, where the names are the new column names
and the values are the expressions to mutate the columns
The name with the suffix \verb{:ident} will be used as the new identity column
The values can be either character strings of expressions to be parsed}

\item{log}{Logger object to log the messages. If NULL, the default logger will be used.}
}
\value{
A Seurat object with mutated metadata
}
\description{
Mutater the Seurat metadata
}
\examples{
\donttest{
obj <- MutateSeuratMeta(
    SeuratObject::pbmc_small,
    list(a = 1, `g:ident` = "paste0(groups, '_ident')")
)
head(obj@meta.data[, c("a", "g")])
GetIdentityColumn(obj)
}
}
