% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-celltype-annotation.R
\name{patch_garnett_make_predictions}
\alias{patch_garnett_make_predictions}
\title{Patch garnett's make_predictions for glmnet >= 4.0 multinomial fits}
\usage{
patch_garnett_make_predictions(log)
}
\arguments{
\item{log}{Logger}
}
\value{
Invisibly, the patched namespace function
}
\description{
Patch garnett's make_predictions for glmnet >= 4.0 multinomial fits
}
