% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-load-qc.R
\name{PerformSeuratCellQC}
\alias{PerformSeuratCellQC}
\title{Perform cell QC}
\usage{
PerformSeuratCellQC(object, cell_qc)
}
\arguments{
\item{object}{Seurat object}

\item{cell_qc}{Cell QC criteria.
It is an expression string to pass to \code{dplyr::filter} function to filter the cells.
It can also be a list of expressions, where the names of the list are sample names.
You can have a default expression in the list with the name "DEFAULT" for the samples
that are not listed.}
}
\value{
The Seurat object with cell QC results in \verb{@misc$cell_qc_df}
}
\description{
Perform cell QC
}
