% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-load-qc.R
\name{PerformSeuratQC}
\alias{PerformSeuratQC}
\title{Perform cell and gene QC on a Seurat object}
\usage{
PerformSeuratQC(object, cell_qc = NULL, gene_qc = NULL)
}
\arguments{
\item{object}{Seurat object of a single sample}

\item{cell_qc}{Cell QC criteria.
It is an expression string to pass to \code{dplyr::filter} function to filter the cells.
It can also be a list of expressions, where the names of the list are sample names.
You can have a default expression in the list with the name "DEFAULT" for the samples
that are not listed.}

\item{gene_qc}{Gene QC criteria
A list containing the following fields:
\itemize{
\item min_cells: Minimum number of cells a gene should be expressed in to be kept
\item excludes: A string or strings to exclude certain genes. Regular expressions are supported.
Multiple strings can also be separated by commas in a single string.
}}
}
\value{
The Seurat object with the \code{.QC} column in \code{meta.data} and, if \code{gene_qc} is
given, the gene QC results in \verb{@misc$gene_qc}
}
\description{
This is a convenience wrapper of \code{\link[=PerformSeuratCellQC]{PerformSeuratCellQC()}} and \code{\link[=PerformGeneQC]{PerformGeneQC()}}
for a single sample, as done by \code{\link[=LoadSeuratAndPerformQC]{LoadSeuratAndPerformQC()}} per sample before
the samples are merged.
}
\examples{
\donttest{
obj <- PerformSeuratQC(
    SeuratObject::pbmc_small,
    cell_qc = "nFeature_RNA > 40",
    gene_qc = list(min_cells = 3)
)
table(obj$.QC)
head(obj@misc$gene_qc)
}
}
