% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{PreRank}
\alias{PreRank}
\title{Pre-rank genes based on expression data}
\usage{
PreRank(
  exprs,
  case,
  control,
  classes,
  method = c("signal_to_noise", "abs_signal_to_noise", "t_test", "ratio_of_classes",
    "diff_of_classes", "log2_ratio_of_classes", "s2n", "abs_s2n")
)
}
\arguments{
\item{exprs}{Expression data matrix (genes x samples)}

\item{case}{The case group name}

\item{control}{The control group name}

\item{classes}{A vector of class labels for each sample
Must be in the same order as the columns of \code{exprs}}

\item{method}{The method to use for ranking
One of "signal_to_noise", "abs_signal_to_noise", "t_test",
"ratio_of_classes", "diff_of_classes", "log2_ratio_of_classes",
"s2n", "abs_s2n"}
}
\value{
A vector of the rank values with names as the gene names
}
\description{
Pre-rank genes based on expression data
}
\seealso{
https://gseapy.readthedocs.io/en/latest/run.html#gseapy.gsea
}
