% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{read_table}
\alias{read_table}
\title{Read a table, like read.table, but with annotated factor levels}
\usage{
read_table(file, factor_level_sep = "|", ...)
}
\arguments{
\item{file}{The file to read}

\item{factor_level_sep}{The separator for factor levels, default is "|"}

\item{...}{Additional arguments passed to read.table}
}
\value{
A data frame with annotated factor levels
}
\description{
The factor levels are annotated like:
\verb{# factor-levels: group=A|B|C}
Column names are used literally, so \code{check.names} defaults to \code{FALSE}.
}
