% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{Reporter}
\alias{Reporter}
\title{Add and save report}
\description{
Add report to a report.json to pipen-report to build a report
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{report}}{The report}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Reporter-new}{\code{Reporter$new()}}
\item \href{#method-Reporter-add}{\code{Reporter$add()}}
\item \href{#method-Reporter-save}{\code{Reporter$save()}}
\item \href{#method-Reporter-clone}{\code{Reporter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reporter-new"></a>}}
\if{latex}{\out{\hypertarget{method-Reporter-new}{}}}
\subsection{Method \code{new()}}{
Initialize the reporter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reporter-add"></a>}}
\if{latex}{\out{\hypertarget{method-Reporter-add}{}}}
\subsection{Method \code{add()}}{
Add a content to the report
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$add(..., h1, h2 = "#", h3 = "#", ui = "flat")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{The content to add}

\item{\code{h1}}{The first level heading of the report}

\item{\code{h2}}{The second level heading of the report}

\item{\code{h3}}{The third level heading of the report}

\item{\code{ui}}{The user interface of the report}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reporter-save"></a>}}
\if{latex}{\out{\hypertarget{method-Reporter-save}{}}}
\subsection{Method \code{save()}}{
Save the report to a file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$save(path, clear = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{The path to save the report
If the path is a directory, the report will be saved as \code{report.json} in the directory
Otherwise, the report will be saved to the path}

\item{\code{clear}}{Whether to clear the report after saving}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Reporter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Reporter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Reporter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
