% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repr.R
\name{repr}
\alias{repr}
\title{The string representation of an object}
\usage{
repr(x, newline = FALSE, ...)
}
\arguments{
\item{x}{An object}

\item{newline}{Whether to add newlines to the output for each element}

\item{...}{Other arguments passed to the \code{.repr} API.}
}
\value{
The string representation
}
\description{
The string representation of an object
}
\examples{
repr(1:10)
# [1] "c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10)"
}
