% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calling.R
\name{run_command}
\alias{run_command}
\title{Run a command}
\usage{
run_command(
  cmd,
  fg = FALSE,
  wait = TRUE,
  print_command = cat,
  stdout = "",
  stderr = "",
  ...
)
}
\arguments{
\item{cmd}{A command to run}

\item{fg}{Whether to run the command in the foreground}

\item{wait}{Whether to wait for the command to finish}

\item{print_command}{The handler to print the command}

\item{stdout}{See \link[base:system2]{base::system2}}

\item{stderr}{See \link[base:system2]{base::system2}}

\item{...}{Additional arguments to pass to \code{system2}}
}
\value{
The output of the command
If \code{stdout} or \code{stderr} is \code{TRUE}, the output is returned as a character vector
Otherwise, the output is returned as an integer exit code
}
\description{
Run a command
}
