% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-load-qc.R
\name{RunContamCorrection}
\alias{RunContamCorrection}
\title{Run contaminant RNA correction on a Seurat object}
\usage{
RunContamCorrection(
  object,
  method,
  decontXArgs = list(),
  scCDCArgs = list(Detection = list(), Quantification = list(), Correction = list()),
  keep_contam_assay = FALSE,
  log = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{method}{Method to use for contaminant RNA correction. Must be one of "decontx"
(using decontX function from the celda package) and "sccdc" (using scCDC package),
case-insensitively.}

\item{decontXArgs}{Arguments to pass to decontX function from the celda package. See \code{?celda::decontX} for details.}

\item{scCDCArgs}{Arguments to pass to scCDC function from the scCDC package.
It is a list with 3 elements: Detection, Quantification and Correction, which are lists of arguments to pass to the corresponding functions
from the scCDC package: \code{scCDC::ContaminationDetection}, \code{scCDC::ContaminationQuantification} and \code{scCDC::ContaminationCorrection}.}

\item{keep_contam_assay}{Whether to keep the \code{Contaminated} assay (the original counts
before contamination correction) in the object. If \code{FALSE} (default), the assay is dropped
right after the correction to save memory.}

\item{log}{Logger}
}
\value{
A Seurat object with contaminant RNA corrected counts in the "RNA" assay, the original
counts in the "Contaminated" assay (unless \code{keep_contam_assay} is \code{FALSE}), and the tool used
in \verb{@misc$contamination$tool}. For scCDC, the detected GCGs and contamination ratios are
recorded in \verb{@misc$contamination} as well.
}
\description{
Run contaminant RNA correction on a Seurat object
}
