% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-macros.R
\name{RunDEAnalysis}
\alias{RunDEAnalysis}
\title{RunDEAnalysis}
\usage{
RunDEAnalysis(
  object,
  group.by,
  ident.1 = NULL,
  ident.2 = NULL,
  subset = NULL,
  cache = NULL,
  error = TRUE,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{group.by}{Column name in meta.data to group cells by}

\item{ident.1}{Identity of the first group of cells}

\item{ident.2}{Identity of the second group of cells}

\item{subset}{Subset of cells to use for analysis
It should be a string of expression to pass to \code{dplyr::filter} function}

\item{cache}{Directory to cache results
The results of the analysis will be saved in this directory
Anything changed in the object or arguments will trigger a re-run}

\item{error}{Whether to raise an error if the analysis fails
Otherwise, return an empty data frame}

\item{...}{Additional arguments to pass to \link[Seurat:FindMarkers]{Seurat::FindMarkers}}
}
\description{
Run differential expression analysis on a Seurat object
}
\examples{
RunDEAnalysis(SeuratObject::pbmc_small, "groups", "g1", "g2")
}
