% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{RunDEGAnalysis}
\alias{RunDEGAnalysis}
\title{Run differential gene expression analysis}
\usage{
RunDEGAnalysis(
  exprs,
  group_by,
  ident_1,
  ident_2 = NULL,
  paired_by = NULL,
  meta = "@meta",
  tool = c("DESeq2", "edgeR", "deseq2", "edger"),
  log = get_logger()
)
}
\arguments{
\item{exprs}{Expression matrix with genes as rows and samples as columns.}

\item{group_by}{Column name in \code{meta} to group samples for differential expression analysis}

\item{ident_1}{First identity to compare against}

\item{ident_2}{Second identity to compare against
If not specified, the rest of the samples will be used as the second identity.}

\item{paired_by}{Column name in \code{meta} for paired samples.
For example, \code{Subject}, for each subject, there should be only one sample in each group (\code{ident_1} and \code{ident_2}).}

\item{meta}{Metadata data frame with sample information. If a \code{Sample} column is present, it will be used as the sample identifier,
which is the column name in \code{exprs}.}

\item{tool}{Tool to use for differential expression analysis.
Currently supports "DESeq2", "edgeR", and "limma".}

\item{log}{Logger}
}
\value{
A data frame with differential expression results.
With attributes:
\itemize{
\item \code{object}: The input expression matrix
\item \code{meta}: The metadata used for the analysis
\item \code{paired_by}: The column name used for paired samples, if applicable.
\item \code{group_by}: The column name used for grouping samples.
\item \code{ident_1}: The first identity used for comparison.
\item \code{ident_2}: The second identity used for comparison.
}
}
\description{
Run differential gene expression analysis
}
