% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{RunEnrichment}
\alias{RunEnrichment}
\title{Run Enrichment Analysis (Over-representation Analysis)}
\usage{
RunEnrichment(
  degs,
  deg = "p_val_adj < 0.05",
  dbs = "KEGG_2021_Human",
  cache = NULL,
  error = TRUE,
  site = "Enrichr",
  ...
)
}
\arguments{
\item{degs}{A data frame of DEGs, typically generated by \code{RunDEAnalysis}}

\item{deg}{A string specifying the condition for a gene to be considered as differentially expressed}

\item{dbs}{A string specifying the database to use for enrichment analysis
A list of available databases can be found at \url{https://maayanlab.cloud/Enrichr/#libraries}}

\item{cache}{A string specifying the directory to cache the results
Set to \code{FALSE} to disable caching}

\item{error}{A boolean specifying whether to throw an error if the enrichment analysis fails}

\item{site}{A string specifying the Enrichr site to use}

\item{...}{Additional arguments to pass to \link{RunEnrichment}}
}
\description{
Run enrichment analysis on a list of differentially expressed genes (DEGs)
}
\examples{
degs <- suppressWarnings(RunDEAnalysis(SeuratObject::pbmc_small, "groups", "g1", "g2"))
RunEnrichment(degs, deg = "abs(avg_log2FC) > 1")
}
