% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{RunGSEA}
\alias{RunGSEA}
\title{Run GSEA using fgsea}
\usage{
RunGSEA(ranks, genesets, ...)
}
\arguments{
\item{ranks}{A named numeric vector of gene ranks}

\item{genesets}{A list of gene sets}

\item{...}{Additional arguments passed to \code{\link[fgsea:fgsea]{fgsea::fgsea()}}}
}
\value{
A data frame with the results of the fGSEA analysis
}
\description{
Run GSEA using fgsea
}
\examples{
\dontrun{
exprs <- matrix(rnorm(1000), nrow = 100, ncol = 10)
colnames(exprs) <- paste0("Sample", 1:10)
rownames(exprs) <- paste0("Gene", 1:100)
classes <- c(rep("A", 5), rep("B", 5))
ranks <- RunGSEAPreRank(exprs, case = "A", control = "B", classes = classes)
genesets <- list(
    set1 = c("Gene1", "Gene2", "Gene3"),
    set2 = c("Gene4", "Gene5", "Gene6")
)
RunGSEA(ranks, genesets)
}
}
