% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-load-qc.R
\name{RunSeuratContamCorrection}
\alias{RunSeuratContamCorrection}
\title{Run contaminant RNA correction on a Seurat object}
\usage{
RunSeuratContamCorrection(
  object,
  method = "decontx",
  decontXArgs = list(),
  scCDCArgs = list(Detection = list(), Quantification = list(), Correction = list())
)
}
\arguments{
\item{object}{Seurat object}

\item{method}{Method to use for contaminant RNA correction.
Supported methods: "decontx" (using decontX function from the celda package) and "sccdc" (using scCDC package).}

\item{decontXArgs}{Arguments to pass to decontX functionn from the celda package. See \code{?celda::decontX} for details.}

\item{scCDCArgs}{Arguments to pass to scCDC function from the scCDC package.
It is a list with 3 elements: Detection, Quantification and Correction, which are lists of arguments to pass to the corresponding functions
from the scCDC package: \code{scCDC::ContaminationDetection}, \code{scCDC::ContaminationQuantification} and \code{scCDC::ContaminationCorrection}.}
}
\value{
A Seurat object with contaminant RNA corrected counts in the "RNA" assay and the original counts in the "Contaminated" assay.
}
\description{
Run contaminant RNA correction on a Seurat object
}
