% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-load-qc.R
\name{RunSeuratDoubletDetection}
\alias{RunSeuratDoubletDetection}
\title{Run doublet detection on a Seurat object}
\usage{
RunSeuratDoubletDetection(
  object,
  tool = "DoubletFinder",
  DoubletFinderArgs = list(),
  scDblFinderArgs = list(),
  filter = TRUE,
  log = NULL,
  cache = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{tool}{Doublet detection tool.
Either "DoubletFinder" or "scDblFinder"}

\item{DoubletFinderArgs}{Arguments to pass to \link{RunSeuratDoubletFinder}}

\item{scDblFinderArgs}{Arguments to pass to \link{RunSeuratScDblFinder}}

\item{filter}{Whether to filter out doublets. Default is TRUE.
If you want to visualize the singlets and doublets, set it to FALSE so that the doublets are not filtered out.
Then you will need to filter them out manually. For example: \code{object <- subset(object, subset = DoubletFinder_DropletType != "doublet")}}

\item{log}{Logger}

\item{cache}{Directory to cache the results. Set to \code{FALSE} to disable caching}
}
\value{
The Seurat object with doublet detection results in \verb{@misc$doublets}
}
\description{
Run doublet detection on a Seurat object
}
