% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-load-qc.R
\name{RunSeuratDoubletFinder}
\alias{RunSeuratDoubletFinder}
\title{Run DoubletFinder on a Seurat object}
\usage{
RunSeuratDoubletFinder(
  object,
  ident = NULL,
  ncores = 1,
  PCs = 30,
  pN = 0.25,
  doublets = 0.075,
  reuse.pANN = NULL,
  log = NULL,
  allow_warnings = FALSE
)
}
\arguments{
\item{object}{Seurat object}

\item{ident}{Cluster identity to use for homotypic doublet proportion estimation.
If NULL or not exists, will run clustering first to get cluster identities.}

\item{ncores}{Number of cores to use}

\item{PCs}{Number of statistically-significant principal components}

\item{pN}{The number of generated artificial doublets, expressed as a proportion of the merged
real-artificial data. Default is set to 0.25, based on observation that DoubletFinder
performance is largely pN-invariant (see McGinnis, Murrow and Gartner 2019, Cell Systems).}

\item{doublets}{The expected proportion of doublets in the dataset. Default is set to 0.075}

\item{reuse.pANN}{Metadata column name for previously-generated pANN results.
If NULL (default), an existing pANN column matching the current pN/pK is
reused to skip pANN computation, if available. Set to FALSE or an invalid
column name is treated as NULL with a warning.}
}
\value{
The Seurat object with doublet detection results in \verb{@misc$doubletFinder}
}
\description{
Run DoubletFinder on a Seurat object
}
\keyword{internal}
