% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-interop.R
\name{RunSeuratMap2Ref}
\alias{RunSeuratMap2Ref}
\title{Run Seurat MapQuery to reference}
\usage{
RunSeuratMap2Ref(
  object,
  ref,
  use,
  ident = "seurat_clusters",
  refnorm = c("auto", "LogNormalize", "SCT", "SCTransform"),
  skip_if_normalized = TRUE,
  split_by = NULL,
  ncores = 1,
  MapQueryArgs = list(refdata = list()),
  FindTransferAnchorsArgs = list(),
  SCTransformArgs = list(),
  NormalizeDataArgs = list(),
  log = NULL,
  cache = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{ref}{Reference Seurat object or a file path to a Seurat object with .rds or .h5seurat extension}

\item{use}{The name in the metadata of the reference to use as identity for the query after mapping}

\item{ident}{The name of the identity in metadata after mapping}

\item{refnorm}{The normalization method used by the reference.
The same normalization method will be used for the query.
\itemize{
\item "auto": automatically detect the normalization method used by the reference
\item "LogNormalize": LogNormalize
\item "SCT": SCTransform
\item "SCTransform": SCTransform (alias for SCT)
}}

\item{skip_if_normalized}{For LogNormalize, skip \code{\link[Seurat:NormalizeData]{Seurat::NormalizeData()}} if the query's default
assay is already 'RNA' (i.e., already log-normalized).
For SCT with \code{FindTransferAnchorsArgs$recompute.residuals = FALSE}, skip \code{\link[Seurat:SCTransform]{Seurat::SCTransform()}} if
the default assay is already 'SCT'.
This parameter has no effect in the default SCT case (\code{recompute.residuals = TRUE}), where
\code{\link[Seurat:FindTransferAnchors]{Seurat::FindTransferAnchors()}} handles normalization internally using the reference model.}

\item{split_by}{The name of the metadata to split the query object by and do the mapping separately}

\item{ncores}{Number of cores to use for parallel processing for the split query objects}

\item{MapQueryArgs}{Arguments to pass to \code{\link[Seurat:MapQuery]{Seurat::MapQuery()}}.
The \code{use} argument will be added to the \code{refdata} list.}

\item{FindTransferAnchorsArgs}{Arguments to pass to \code{\link[Seurat:FindTransferAnchors]{Seurat::FindTransferAnchors()}}.}

\item{SCTransformArgs}{Arguments to pass to \code{\link[Seurat:SCTransform]{Seurat::SCTransform()}}.
Only used when \code{refnorm} is set to "SCT" AND \code{FindTransferAnchorsArgs$recompute.residuals} is
explicitly set to \code{FALSE}. In the default case (\code{recompute.residuals = TRUE}),
\code{\link[Seurat:FindTransferAnchors]{Seurat::FindTransferAnchors()}} recomputes query residuals from raw counts using the reference
SCT model, making manual \code{\link[Seurat:SCTransform]{Seurat::SCTransform()}} on the query unnecessary.}

\item{NormalizeDataArgs}{Arguments to pass to \code{\link[Seurat:NormalizeData]{Seurat::NormalizeData()}}.
Will be used to normalize the query object if \code{refnorm} is set to "LogNormalize"}

\item{log}{Logger}

\item{cache}{Directory to cache the results. Set to \code{FALSE} to disable caching}
}
\value{
The Seurat object with the mapped data
}
\description{
This will find transfer anchors between query and reference, and map the query to the reference.
For SCT-normalized references, \code{\link[Seurat:FindTransferAnchors]{Seurat::FindTransferAnchors()}} handles query normalization internally
by recomputing residuals using the reference's SCT model (\code{recompute.residuals = TRUE} by default),
so a manual call to \code{\link[Seurat:SCTransform]{Seurat::SCTransform()}} on the query is NOT needed.
For LogNormalize references, the query is normalized with \code{\link[Seurat:NormalizeData]{Seurat::NormalizeData()}} prior to anchor
finding.
}
\seealso{
https://satijalab.org/seurat/articles/integration_mapping.html

https://satijalab.org/seurat/articles/multimodal_reference_mapping.html

https://satijalab.org/seurat/articles/covid_sctmapping
}
