% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-clustering.R
\name{RunSeuratPreparation}
\alias{RunSeuratPreparation}
\title{Prepare a Seurat object for downstream analysis}
\usage{
RunSeuratPreparation(
  object,
  use_sct = FALSE,
  SCTransformArgs = list(),
  NormalizeDataArgs = list(),
  FindVariableFeaturesArgs = list(),
  ScaleDataArgs = list(),
  RunPCAArgs = list(),
  no_integration = FALSE,
  IntegrateLayersArgs = list(),
  log = NULL,
  cache = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{use_sct}{Whether to use \link[Seurat:SCTransform]{Seurat::SCTransform}}

\item{SCTransformArgs}{Arguments to pass to \link[Seurat:SCTransform]{Seurat::SCTransform}}

\item{NormalizeDataArgs}{Arguments to pass to \link[Seurat:NormalizeData]{Seurat::NormalizeData}}

\item{FindVariableFeaturesArgs}{Arguments to pass to \link[Seurat:FindVariableFeatures]{Seurat::FindVariableFeatures}}

\item{ScaleDataArgs}{Arguments to pass to \link[Seurat:ScaleData]{Seurat::ScaleData}
You can use \code{features = "__all__"} to scale all features, which will be expanded to all features in the object.}

\item{RunPCAArgs}{Arguments to pass to \link[Seurat:RunPCA]{Seurat::RunPCA}}

\item{no_integration}{Whether to skip integration, just join layers}

\item{IntegrateLayersArgs}{Arguments to pass to \link[Seurat:IntegrateLayers]{Seurat::IntegrateLayers}}

\item{log}{Logger}

\item{cache}{Directory to cache the results. Set to \code{FALSE} to disable caching}
}
\value{
The transformed and integrated Seurat object
}
\description{
Data transformation (via \code{\link[=RunSeuratTransformation]{RunSeuratTransformation()}}) and sample integration
(via \code{\link[=RunSeuratIntegration]{RunSeuratIntegration()}}) are run in one go.
}
\examples{
\donttest{
# `pbmc_small` ships with a v3 assay, which integration doesn't support
obj <- SeuratObject::pbmc_small
obj[["RNA"]] <- as(obj[["RNA"]], "Assay5")
obj <- RunSeuratPreparation(obj, no_integration = TRUE)
}
}
