% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{save_plot}
\alias{save_plot}
\title{Save the plot into multiple formats}
\usage{
save_plot(
  plot,
  prefix,
  devpars = NULL,
  bg = "white",
  formats = c("png", "pdf")
)
}
\arguments{
\item{plot}{The plot object}

\item{prefix}{The prefix of the file
The ending \code{.} is removed from the prefix. So both \verb{/path/to/file} and \verb{/path/to/file.} are valid and will
save the files as \verb{/path/to/file.png}, \verb{/path/to/file.pdf} and \verb{/path/to/file.<format>} etc.}

\item{devpars}{The device parameters}

\item{bg}{The background color}

\item{formats}{The formats to save}
}
\description{
Save the plot into multiple formats
}
