% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{save_plotcode}
\alias{save_plotcode}
\title{Save the plot code and data to reproduce the plot}
\usage{
save_plotcode(plot, prefix, setup = NULL, ..., envir = parent.frame())
}
\arguments{
\item{plot}{The plot object}

\item{prefix}{The prefix of the file
The ending \code{.} is removed from the prefix. So both \verb{/path/to/file} and \verb{/path/to/file.} are valid and will
save the files as \verb{/path/to/file.code.zip}.}

\item{setup}{The code to setup the environment for the plot generation}

\item{...}{The names of the variables to save from the environment}

\item{envir}{The environment to save the variables from}
}
\description{
The plot should be generated by functions that registered by \link[gglogger:register]{gglogger::register}
}
