% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{save_to_cache}
\alias{save_to_cache}
\title{Save an object to cache}
\usage{
save_to_cache(to_cache, kind, cache_dir)
}
\arguments{
\item{to_cache}{An list to cache,
including the signature, digested signature and data}

\item{kind}{A string indicating the kind of the object
Used as part of the filename of the cached file}

\item{cache_dir}{A string indicating the directory to store cached files}
}
\description{
Save an object to cache
}
