% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctcr-marcos.R
\name{ScRepSubset}
\alias{ScRepSubset}
\title{Subset scRepertorie object}
\usage{
ScRepSubset(screp, subset)
}
\arguments{
\item{screp}{The scRepertorie object. It is either a Seurat object or a list of data.frames}

\item{subset}{The subset expression (in characters)}
}
\value{
The subsetted scRepertorie object
}
\description{
Subset scRepertorie object
}
\examples{
\donttest{
data(contig_list, package = "scRepertoire")
screp <- scRepertoire::combineTCR(contig_list,
   samples = c("P17B", "P17L", "P18B", "P18L", "P19B","P19L", "P20B", "P20L"))

head(ScRepSubset(screp, "nchar(CTaa) < 20")[[1]])
names(ScRepSubset(screp, "Sample \%in\% c('P17B', 'P17L')"))
}
}
