% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq.R
\name{VizBulkDEGs}
\alias{VizBulkDEGs}
\title{Visualize DEGs}
\usage{
VizBulkDEGs(
  degs,
  plot_type = c("volcano", "violin", "box", "bar", "ridge", "heatmap", "cor"),
  order_by = "desc(abs(log2FC))",
  genes = 10,
  outprefix = NULL,
  devpars = list(res = 100),
  more_formats = c(),
  save_code = FALSE,
  show_row_names = TRUE,
  show_column_names = TRUE,
  ...
)
}
\arguments{
\item{degs}{DEGs from RunDEGAnalysis}

\item{plot_type}{Type of plot to generate
One of 'volcano', 'violin', 'box', 'bar', 'ridge', 'dim', 'heatmap', 'dot'}

\item{order_by}{An expression in string to order the genes}

\item{genes}{Number of genes genes to visualize (based on the 'order_by' expression)
Or an expression in string to filter the genes (passed by \link[dplyr:filter]{dplyr::filter})
Only works when plot_type is not a volcano plot}

\item{outprefix}{Prefix of the output file}

\item{devpars}{List of parameters to save the plot}

\item{more_formats}{Additional formats to save the plot in addition to 'png'}

\item{save_code}{Whether to save the code to reproduce the plot}

\item{show_row_names}{Whether to show row names in the heatmap}

\item{show_column_names}{Whether to show column names in the heatmap}

\item{...}{Additional arguments to pass to the plot function
\itemize{
\item For 'volcano', additional arguments to pass to 'scplotter::VolcanoPlot'
\item For 'violin', 'box', 'bar', 'ridge', 'dim', 'heatmap', 'dot', additional arguments to pass to 'scplotter::FeatureStatPlot'
}}
}
\value{
A ggplot object if 'outprefix' is NULL, otherwise, save the plot to the output directory
}
\description{
Visualize bulk differentially expressed genes
}
\examples{
\donttest{
set.seed(8525)
data = matrix(rnbinom(1000, mu = 5, size = 1), nrow = 100, ncol = 10)
rownames(data) <- paste0("Gene", 1:100)
colnames(data) <- paste0("Sample", 1:10)
meta <- data.frame(
   Sample = colnames(data),
   Condition = rep(c("Control", "Treatment"), each = 5)
)

degs <- RunDEGAnalysis(data, meta = meta,
 group_by = "Condition", ident_1 = "Treatment", tool = "edger")
VizBulkDEGs(degs, plot_type = "volcano", legend.position = "top", y_cutoff = 0.25)
VizBulkDEGs(degs, plot_type = "violin", genes = 2)
VizBulkDEGs(degs, plot_type = "violin", stack = TRUE, genes = 2)
VizBulkDEGs(degs, plot_type = "box", genes = 2)
VizBulkDEGs(degs, plot_type = "bar", genes = 2, x_text_angle = 90)
VizBulkDEGs(degs, plot_type = "ridge", genes = 2)
VizBulkDEGs(degs, plot_type = "heatmap", genes = 5)
}
}
