% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsea.R
\name{VizEnrich}
\alias{VizEnrich}
\title{Visualize Enrichment}
\usage{
VizEnrich(
  enrich,
  outprefix = NULL,
  devpars = list(res = 100),
  more_formats = c(),
  save_code = FALSE,
  ...
)
}
\arguments{
\item{enrich}{Enrichment object from RunEnrichment}

\item{outprefix}{Prefix of the output file}

\item{devpars}{List of parameters to save the plot}

\item{more_formats}{Additional formats to save the plot in addition to 'png'}

\item{save_code}{Whether to save the code to reproduce the plot}

\item{...}{Additional arguments to pass to the plot function \link[scplotter:EnrichmentPlot]{scplotter::EnrichmentPlot}}
}
\value{
A ggplot object if 'outprefix' is NULL, otherwise, save the plot to the output directory
}
\description{
Visualize enrichment (over-representation) analysis results
}
\examples{
\donttest{
degs <- suppressWarnings(RunDEAnalysis(SeuratObject::pbmc_small, "groups", "g1", "g2"))
enrich <- RunEnrichment(degs, deg = "abs(avg_log2FC) > 1")
VizEnrich(enrich)
VizEnrich(enrich, plot_type = "dot")
VizEnrich(enrich, plot_type = "lollipop")
VizEnrich(enrich, plot_type = "network")
# VizEnrich(enrich, plot_type = "enrichmap")
# VizEnrich(enrich, plot_type = "wordcloud")
}
}
