% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-viz.R
\name{VizSeuratCellQC}
\alias{VizSeuratCellQC}
\title{Visualize Cell QC metrics of Seurat object}
\usage{
VizSeuratCellQC(
  object,
  features = c("nFeature_RNA", "nCount_RNA", "percent.mt", "percent.ribo", "percent.hb",
    "percent.plat"),
  plot_type = c("violin", "box", "scatter", "ridge"),
  scatter_x = "nCount_RNA",
  palette = "Set2",
  ...
)
}
\arguments{
\item{object}{A Seurat object with cell QC metrics}

\item{features}{Features to visualize}

\item{plot_type}{Type of plot to generate
One of 'violin', 'box', 'scatter', 'ridge'}

\item{scatter_x}{Feature to use as x-axis in scatter plot.
If it is one of the features, it will be removed from the features}

\item{palette}{Color palette to use}

\item{...}{Additional arguments to pass to the plot function}
}
\value{
The plot
}
\description{
Visualize Cell QC metrics of Seurat object
}
