% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-viz.R
\name{VizSeuratContamination}
\alias{VizSeuratContamination}
\title{Visualize contamination correction results of Seurat object}
\usage{
VizSeuratContamination(
  object,
  plot_type = c("ridge", "histogram", "bar", "violin", "box", "heatmap", "dot"),
  metric = c("contam", "contamination", "expr", "expression"),
  markers = NULL,
  ...
)
}
\arguments{
\item{object}{A Seurat object with contamination correction results}

\item{plot_type}{Type of plot to generate}

\item{metric}{Metric to visualize, one of "contam", "contamination" and "expression". "contam" is an alias of "contamination".
For \code{scCDC}, only sample-level contamination fraction is available.
For \code{decontX}, cell-level contamination fraction is available.}

\item{markers}{List of marker genes to visualize expression of contamination genes.
For \code{scCDC}, the \code{GCGs} will be used by default.}

\item{...}{Additional arguments to pass to the plot function.}
}
\description{
Visualize contamination correction results of Seurat object
}
