% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-viz.R
\name{VizSeuratDoublets}
\alias{VizSeuratDoublets}
\title{Visualize detected doublets}
\usage{
VizSeuratDoublets(
  object,
  plot_type = c("dim", "pie", "pk", "pK"),
  palette = "Set2",
  ...
)
}
\arguments{
\item{object}{A Seurat object with detected doublets}

\item{plot_type}{Type of plot to generate
One of 'dim', 'pie', 'pk', 'pK'.
dim/pie show the distribution of doublets by droplet type.
pk/pK show the relationship between BC metric and pK when using DoubletFinder.}

\item{palette}{Color palette to use}

\item{...}{Additional arguments to pass to the plot function
\itemize{
\item For 'dim',  additional arguments to pass to \link[scplotter:CellDimPlot]{scplotter::CellDimPlot}
\item For 'pie',  additional arguments to pass to \link[plotthis:PieChart]{plotthis::PieChart}
\item For 'pk' or 'pK', additional arguments to pass to \link[plotthis:LinePlot]{plotthis::LinePlot}
}}
}
\value{
The plot
}
\description{
Visualize detected doublets
}
\examples{
\donttest{
datadir <- system.file("extdata", "scrna", package = "biopipen.utils")
meta <- data.frame(
    Sample = c("Sample1", "Sample2"),
    RNAData = c(
        file.path(datadir, "Sample1"),
        file.path(datadir, "Sample2")
    )
)

obj <- LoadSeuratAndPerformQC(meta, cache = FALSE, gene_qc = list(min_cells = 3),
    cell_qc = "nFeature_RNA > 500")
VizSeuratCellQC(obj)
VizSeuratCellQC(obj, plot_type = "scatter")
VizSeuratCellQC(obj, plot_type = "ridge")
VizSeuratGeneQC(obj)

obj <- RunSeuratTransformation(obj)
obj <- RunSeuratIntegration(obj)
obj <- RunSeuratDoubletDetection(obj, tool = "scDblFinder", filter = FALSE)
VizSeuratDoublets(obj)
VizSeuratDoublets(obj, plot_type = "pie")
}
}
