% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-viz.R
\name{VizSeuratGeneQC}
\alias{VizSeuratGeneQC}
\title{Visualize gene QC metrics of Seurat object}
\usage{
VizSeuratGeneQC(object, features = 10, ...)
}
\arguments{
\item{object}{A Seurat object with gene QC metrics}

\item{features}{Number of features/actual features to visualize
If actual features are given, they must be the features excluded in gene QC}

\item{...}{Additional arguments to pass to the plot function \link[plotthis:barplot]{plotthis::BarPlot}}
}
\value{
The plot
}
\description{
Visualize gene QC metrics of Seurat object
}
