% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{write_table}
\alias{write_table}
\title{Write a table, like write.table, but with annotated factor levels}
\usage{
write_table(x, file, factor_level_sep = "|", ...)
}
\arguments{
\item{x}{The data frame to write}

\item{file}{The file to write}

\item{factor_level_sep}{The separator for factor levels, default is "|"}

\item{...}{Additional arguments passed to write.table}
}
\description{
The factor levels are annotated like:
\verb{# factor-levels: group=A|B|C}
}
