% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat-disk-patch.R
\name{WriteH5Group.Seurat}
\alias{WriteH5Group.Seurat}
\title{Patch to fix SeuratDisk::WriteH5Group for Assay objects
SeuratObject 5.3.0 raises an error with SeuratDisk::WriteH5Group() for Assay objects}
\usage{
WriteH5Group.Seurat(x, name, hgroup, verbose = TRUE)
}
\arguments{
\item{x}{Assay object}

\item{name}{Name of the group to create}

\item{hgroup}{HDF5 group to write to}

\item{verbose}{Whether to print messages}
}
\value{
Invisibly returns \code{NULL}
}
\description{
Patch to fix SeuratDisk::WriteH5Group for Assay objects
SeuratObject 5.3.0 raises an error with SeuratDisk::WriteH5Group() for Assay objects
}
\keyword{internal}
