% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-macros.R
\name{AggregateExpressionPseudobulk}
\alias{AggregateExpressionPseudobulk}
\title{Aggregate expression of single cells into psedobulk expression matrix}
\usage{
AggregateExpressionPseudobulk(
  object,
  aggregate_by = "Sample",
  assay = "RNA",
  layer = "counts",
  log = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{aggregate_by}{The metadata column to aggregate by. Default is "Sample".
You can add other columns to aggregate by, such as "Condition", "Batch", etc. But
they should superset the "Sample" column, meaning that one sample can only belong to one condition,
but one condition can have multiple samples.}

\item{assay}{The assay to use for aggregation. Default is "RNA".}

\item{layer}{The layer to use for aggregation. Default is "counts".}

\item{log}{Logger}
}
\value{
A list of two elements:
\itemize{
\item \code{exprs}: A matrix of aggregated expression values with genes as rows and samples as columns.
\item \code{meta}: A data frame with metadata of the aggregated samples.
}
}
\description{
Aggregate expression of single cells into psedobulk expression matrix
}
\examples{
\donttest{
obj <- SeuratObject::pbmc_small
obj$Sample <- rep(paste0("S", 1:10), each = ncol(obj) / 10)
obj$Condition <- rep(c("Control", "Treatment"), each = ncol(obj) / 2)
result <- AggregateExpressionPseudobulk(obj, aggregate_by = c("Sample", "Condition"))
head(result$exprs)
head(result$meta)
}
}
