% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-macros.R
\name{ConvertSeuratToAnnData}
\alias{ConvertSeuratToAnnData}
\title{Convert a Seurat object (or RDS/H5Seurat file) to an AnnData object file}
\usage{
ConvertSeuratToAnnData(object_or_file, outfile, assay = NULL, log = NULL)
}
\arguments{
\item{object_or_file}{The Seurat object or the path to the RDS or H5Seurat file}

\item{outfile}{Output file}

\item{assay}{Assay to be used}

\item{log}{Logger}
}
\value{
No return value
}
\description{
Convert a Seurat object (or RDS/H5Seurat file) to an AnnData object file
}
\examples{
ConvertSeuratToAnnData(SeuratObject::pbmc_small, "/tmp/pbmc_small.h5ad")

}
