% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat-disk-patch.R
\name{.set_h5_seurat_project<-}
\alias{.set_h5_seurat_project<-}
\title{SeuratDisk's "Project<-.h5Seurat" function}
\usage{
.set_h5_seurat_project(object, ...) <- value
}
\description{
SeuratDisk's "Project<-.h5Seurat" function
}
\keyword{internal}
