% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-macros.R
\name{EnsureSeuratScaleData}
\alias{EnsureSeuratScaleData}
\title{Ensure marker genes are in the scale.data layer of an assay of the Seurat object}
\usage{
EnsureSeuratScaleData(object, features, assay = NULL, umi_assay = "RNA")
}
\arguments{
\item{object}{Seurat object}

\item{features}{Character vector of feature names to ensure in the scale.data layer}

\item{assay}{Assay to use. If NULL, the default assay will be used.}

\item{umi_assay}{Assay to use for the UMI counts. Default is "RNA". This is used to get the counts for scaling.}

\item{log}{Logger object to log the messages. If NULL, the default logger will be used.}
}
\value{
The Seurat object with the features ensured in the scale.data layer
}
\description{
Ensure marker genes are in the scale.data layer of an assay of the Seurat object
}
