% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casing.R
\name{expand_cases}
\alias{expand_cases}
\title{Expand the cases with default values}
\usage{
expand_cases(cases, defaults, expand = NULL, default_case = "DEFAULT")
}
\arguments{
\item{cases}{A list of cases}

\item{defaults}{A list of default values}

\item{expand}{A function to expand each case, if NULL, then the \code{each} key will be ignored.
The function should take two arguments, \code{name} and \code{case}, and return a list of expanded cases.}

\item{default_case}{The name of the default case}
}
\value{
A list of expanded cases
}
\description{
Expand the cases with default values
}
