% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{extract_vars}
\alias{extract_vars}
\title{Extract variables from a named list}
\usage{
extract_vars(x, ..., keep = FALSE, env = parent.frame())
}
\arguments{
\item{x}{A named list}

\item{...}{The names of the variables
named arguments are allowed to rename the variables. \code{b = "a"} will extract \code{a} and assign it to \code{b}}

\item{keep}{Keep the extracted variables in the list}

\item{env}{The environment to assign the extracted variables}
}
\value{
The list with/ithout the extracted variables
}
\description{
Extract variables from a named list
}
