% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-macros.R
\name{FinishSeuratQC}
\alias{FinishSeuratQC}
\title{Finish the QC process including the visualization}
\usage{
FinishSeuratQC(object)
}
\arguments{
\item{object}{Seurat object}
}
\value{
The Seurat object with the QC process finished
}
\description{
This will remove the cells and genes that are not passing the QC, and also remove
the intermediate data used for QC visualization.
}
