% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-macros.R
\name{GetIdentityColumn}
\alias{GetIdentityColumn}
\title{Get the column name in meta.data that works as identity}
\usage{
GetIdentityColumn(object)
}
\arguments{
\item{object}{Seurat object}
}
\value{
The column name in meta.data that works as identity
If there are multiple columns matching, return the shortest one.
}
\description{
Get the column name in meta.data that works as identity
}
\examples{
obj <- SeuratObject::pbmc_small
GetIdentityColumn(obj)

SeuratObject::Idents(obj) <- "groups"
GetIdentityColumn(obj)
}
