% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-macros.R
\name{LoadSeuratAndPerformQC}
\alias{LoadSeuratAndPerformQC}
\title{Load samples into a Seurat object}
\usage{
LoadSeuratAndPerformQC(
  meta,
  min_cells = 0,
  min_features = 0,
  samples = NULL,
  cell_qc = NULL,
  gene_qc = NULL,
  tmpdir = NULL,
  log = NULL,
  cache = NULL
)
}
\arguments{
\item{meta}{Metadata of the samples
Required columns: Sample, RNAData.
The RNAData column should contain the path to the 10X data, either a directory or a file
If the path is a directory, the function will look for barcodes.tsv.gz, features.tsv.gz and matrix.mtx.gz.
The directory should be loaded by \link[Seurat:Read10X]{Seurat::Read10X}. Sometimes, there may be prefix in the file names,
e.g. "'prefix'.barcodes.tsv.gz", which is also supported.
If the path is a file, it should be a h5 file that can be loaded by \code{\link[Seurat:Read10X_h5]{Seurat::Read10X_h5()}}}

\item{min_cells}{Include features detected in at least this many cells.
This will be applied to all samples and passed to the \code{\link[Seurat:reexports]{Seurat::CreateSeuratObject()}} function.
QCs can be further performed on the object after loading.
You can also provide a list of values, where the names of the list are sample names
and the values are the minimum number of cells for each sample to load by
\code{\link[Seurat:reexports]{Seurat::CreateSeuratObject()}}.
You can have a default value in the list with the name "DEFAULT" for the samples
that are not listed.}

\item{min_features}{Include cells where at least this many features are detected.
This will be applied to all samples and passed to the \code{\link[Seurat:reexports]{Seurat::CreateSeuratObject()}} function.
QCs can be further performed on the object after loading.
You can also provide a list of values, where the names of the list are sample names
and the values are the minimum number of features for each sample to load by
\code{\link[Seurat:reexports]{Seurat::CreateSeuratObject()}}.
You can have a default value in the list with the name "DEFAULT" for the samples
that are not listed.}

\item{samples}{Samples to load. If NULL, all samples will be loaded}

\item{cell_qc}{Cell QC criteria, a string of expression to pass to \code{dplyr::filter} function
to filter the cells.
It can also be a list of expressions, where the names of the list are sample names.
You can have a default expression in the list with the name "DEFAULT" for the samples
that are not listed.}

\item{gene_qc}{Gene QC criteria
A list containing the following fields:
\itemize{
\item min_cells: Minimum number of cells a gene should be expressed in to be kept
\item excludes: A string or strings to exclude certain genes. Regular expressions are supported.
Multiple strings can also be separated by commas in a single string.
}}

\item{tmpdir}{Temporary directory to store intermediate files when there are prefix in the file names}

\item{log}{Logger}

\item{cache}{Directory to cache the results. Set to \code{FALSE} to disable caching}
}
\value{
A Seurat object.
In slot \code{misc}, there are two fields: \code{cell_qc_df} and \code{gene_qc}, a list of gene QC results
}
\description{
Cell QC will be performed, either per-sample or on the whole object
}
\examples{
\donttest{
datadir <- system.file("extdata", "scrna", package = "biopipen.utils")
meta <- data.frame(
    Sample = c("Sample1", "Sample2"),
    RNAData = c(
        file.path(datadir, "Sample1"),
        file.path(datadir, "Sample2")
    )
)
obj <- LoadSeuratAndPerformQC(meta, cache = FALSE, gene_qc = list(min_cells = 3))
head(obj@misc$cell_qc_df)
print(obj@misc$gene_qc)
}
}
