% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-macros.R
\name{LoadSeuratAndPerformQC}
\alias{LoadSeuratAndPerformQC}
\title{Load samples into a Seurat object}
\usage{
LoadSeuratAndPerformQC(
  meta,
  samples = NULL,
  per_sample_qc = FALSE,
  cell_qc = NULL,
  gene_qc = NULL,
  tmpdir = NULL,
  log = NULL,
  cache = NULL
)
}
\arguments{
\item{meta}{Metadata of the samples
Required columns: Sample, RNAData.
The RNAData column should contain the path to the 10X data, either a directory or a file
If the path is a directory, the function will look for barcodes.tsv.gz, features.tsv.gz and matrix.mtx.gz.
The directory should be loaded by \link[Seurat:Read10X]{Seurat::Read10X}. Sometimes, there may be prefix in the file names,
e.g. "\if{html}{\out{<prefix>}}.barcodes.tsv.gz", which is also supported.
If the path is a file, it should be a h5 file that can be loaded by \link[Seurat:Read10X_h5]{Seurat::Read10X_h5}}

\item{samples}{Samples to load. If NULL, all samples will be loaded}

\item{per_sample_qc}{Whether to perform per-sample cell QC}

\item{cell_qc}{Cell QC criteria}

\item{gene_qc}{Gene QC criteria
A list containing the following fields:
\itemize{
\item min_cells: Minimum number of cells a gene should be expressed in to be kept
\item excludes: A string or strings to exclude certain genes. Regular expressions are supported.
Multiple strings can also be separated by commas in a single string.
}}

\item{tmpdir}{Temporary directory to store intermediate files when there are prefix in the file names}

\item{log}{Logger}

\item{cache}{Directory to cache the results. Set to \code{FALSE} to disable caching}
}
\value{
A Seurat object.
In slot \code{misc}, there are two fields: \code{cell_qc_df} and \code{gene_qc}, a list of gene QC results
}
\description{
Cell QC will be performed, either per-sample or on the whole object
}
\examples{
\donttest{
datadir <- system.file("extdata", "scrna", package = "biopipen.utils")
meta <- data.frame(
    Sample = c("Sample1", "Sample2"),
    RNAData = c(
        file.path(datadir, "Sample1"),
        file.path(datadir, "Sample2")
    )
)
obj <- LoadSeuratAndPerformQC(meta, cache = FALSE, gene_qc = list(min_cells = 3))
head(obj@misc$cell_qc_df)
print(obj@misc$gene_qc)
}
}
