% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-macros.R
\name{PerformGeneQC}
\alias{PerformGeneQC}
\title{Perform gene QC}
\usage{
PerformGeneQC(object, gene_qc)
}
\arguments{
\item{object}{Seurat object of a single sample.}

\item{gene_qc}{Gene QC criteria}
}
\value{
A data frame with Sample (\code{object@project.name}),
Count (number of cells a gene is expressed in) and QC (whether the gene passes the QC)
}
\description{
Perform gene QC
}
