% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-macros.R
\name{PerformSeuratCellQC}
\alias{PerformSeuratCellQC}
\title{Perform cell QC}
\usage{
PerformSeuratCellQC(object, cell_qc)
}
\arguments{
\item{object}{Seurat object}

\item{cell_qc}{Cell QC criteria}
}
\value{
The Seurat object with cell QC results in \verb{@misc$cell_qc_df}
}
\description{
Perform cell QC
}
