% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{read_obj}
\alias{read_obj}
\alias{load_obj}
\alias{save_obj}
\alias{write_obj}
\title{Read and write objects to/from files}
\usage{
read_obj(file, type = c("auto", "qs2", "rds", "h5seurat", "h5ad"), ...)

load_obj(file, type = c("auto", "qs2", "rds", "h5seurat", "h5ad"), ...)

save_obj(obj, file, type = c("auto", "qs2", "rds", "h5seurat", "h5ad"), ...)

write_obj(obj, file, type = c("auto", "qs2", "rds", "h5seurat", "h5ad"), ...)
}
\arguments{
\item{file}{The file to save the object to}

\item{type}{The type of the file. Can be "auto", "qs2", "rds", "h5seurat" or "h5ad".
If "auto", the type will be inferred from the file extension.}

\item{...}{Additional arguments passed to the underlying save functions.}

\item{obj}{The object to save}
}
\value{
The object read from the file or NULL if writing
}
\description{
Read and write objects to/from files

Alias of read_obj

Save an object to a file

Alias of save_obj
}
