% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{read_obj}
\alias{read_obj}
\alias{save_obj}
\title{Read and write objects to/from files}
\usage{
read_obj(file, type = c("auto", "qs2", "rds"))

save_obj(obj, file, type = c("auto", "qs2", "rds"))
}
\arguments{
\item{file}{The file to read or write}

\item{type}{The type of the file. Can be "auto", "qs2" or "rds".
If "auto", the type will be inferred from the file extension.}
}
\value{
The object read from the file or NULL if writing
}
\description{
Read and write objects to/from files

Save an object to a file
}
