% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-macros.R
\name{RenameSeuratIdents}
\alias{RenameSeuratIdents}
\title{Rename cluster names}
\usage{
RenameSeuratIdents(
  object,
  ident = NULL,
  save_as = NULL,
  merge = FALSE,
  backup = NULL,
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{ident}{Column name in \verb{@meta.data} to rename
Defaults to the identity column (from \code{GetIdentityColumn()})
When \code{ident} is not identical as the column identified by \code{GetIdentityColumn()},}

\item{save_as}{If not NULL, save the renamed clusters to this new column.
If NULL, will overwrite the original \code{ident} column.}

\item{merge}{Whether to merge the clusters with the same new name.
Default is FALSE. Otherwise, the clusters with the same new name will be suffixed
with ".1", ".2", etc. to make them unique.}

\item{backup}{If not NULL, the original \code{ident} column will be backed up
to this new column before renaming.
Default is NULL, which means no backup will be made.
Note that if the column name already exists, it will be overwritten.}

\item{...}{Additional arguments.
If named arguments are provided, they will be used as the mapping from old cluster names to new cluster names.
\itemize{
\item If a named argument's name does not exist in the original cluster names, it will be ignored.
\item If a original cluster name does not have a corresponding named argument, it will be kept unchanged.
\item If you want to exclude some clusters, you can set their new names to \code{NA}.
If unnamed arguments are provided, it must be a single list of mappings.
with the names being the original cluster names, and the values being the new cluster names.
}}
}
\value{
The Seurat object with renamed clusters
}
\description{
Rename cluster names
}
\details{
This function provides a convenient way to rename clusters in a Seurat object.
You can provide the mapping of old cluster names to new cluster names either
as named arguments or as a single list.
Unlike \code{\link[Seurat:reexports]{Seurat::RenameIdents()}}, this function also renames the cluster names in the metadata column,
so that the changes are persistent. If \code{save_as} is provided, the renamed clusters will be saved
to a new column, otherwise the original column will be overwritten.
In addition, the order of the clusters will be preserved based on the order of the original cluster names.
For the default identities (\code{Idents(object)}), it is only changed when \code{ident} is NULL or identical
to the identity column (\code{GetIdentityColumn(object)}).
}
\examples{
\donttest{
object <- scplotter::pancreas_sub
GetIdentityColumn(object)
table(object$seurat_clusters)

# Rename using Seurat::RenameIdents
object1 <- Seurat::RenameIdents(object, `2` = 'Alpha', `1` = 'Beta')
# The order of identities is changed
table(SeuratObject::Idents(object1))
# The identity column is not changed
table(object1$seurat_clusters)
# Since the inconsistency, GetIdentityColumn can't determine the identity column
GetIdentityColumn(object1)

# Rename using RenameSeuratIdents
object1 <- RenameSeuratIdents(object, `2` = 'Alpha', `1` = 'Beta')
# The order of identities is preserved
table(SeuratObject::Idents(object1))
# The identity column is also changed
table(object1$seurat_clusters)
# Since the consistency, GetIdentityColumn can still determine the identity column
GetIdentityColumn(object1)

# Rename a non-identity column
table(object$CellType)
object1 <- RenameSeuratIdents(
    object, Ductal = 'C1', `Endocrine` = 'C1', ident = 'CellType')
table(object1$CellType)
table(object1$seurat_clusters)
# Identity column is not changed
GetIdentityColumn(object1)

# Merge clusters with the same new name
object1 <- RenameSeuratIdents(
   object, Ductal = 'C1', `Endocrine` = 'C1', ident = 'CellType', merge = TRUE)
table(object1$CellType)

# Exclude some clusters by setting their new names to NA
object1 <- RenameSeuratIdents(
   object, Ductal = 'C1', `Endocrine` = NA, ident = 'CellType')
table(object1$CellType)
}
}
