% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-macros.R
\name{RunSeuratAmbientRemoval}
\alias{RunSeuratAmbientRemoval}
\title{Run ambient RNA removal on a Seurat object}
\usage{
RunSeuratAmbientRemoval(object, method = "decontx", decontXArgs = list())
}
\arguments{
\item{object}{Seurat object}

\item{method}{Method to use for ambient RNA removal. Currently only "decontx" is supported.}

\item{decontXArgs}{Arguments to pass to decontX functionn from the celda package. See \code{?celda::decontX} for details.}
}
\value{
A Seurat object with ambient RNA removed
}
\description{
Run ambient RNA removal on a Seurat object
}
