% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-macros.R
\name{RunSeuratCellCycleScoring}
\alias{RunSeuratCellCycleScoring}
\title{Run Seurat CellCycleScoring}
\usage{
RunSeuratCellCycleScoring(
  object,
  s.features = NULL,
  g2m.features = NULL,
  species = "auto",
  log = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{s.features}{The gene names for S phase scoring.
Default is \code{cc.genes$s.genes} for human and the following genes for mouse:
Mcm5, Pcna, Tyms, Fen1, Mcm7, Mcm4, Rrm1, Ung, Gins2, Mcm6, Cdca7, Dtl, Prim1, Uhrf1,
Cenpu, Hells, Rfc2, Polr1b, Nasp, Rad51ap1, Gmnn, Wdr76, Slbp, Ccne2, Ubr7, Msh2, Rad51,
Rrm2, Cdc45, Cdc6, Exo1, Tipin, Dscc1, Blm, Casp8ap2, Usp1, Clspn, Pola1, Chaf1b, Mrpl36, E2f8}

\item{g2m.features}{The gene names for G2/M phase scoring.
Default is \code{cc.genes$g2m.genes} for human and the following genes for mouse:
Hmgb2, Cdk1, Nusap1, Ube2c, Birc5, Tpx2, Top2a, Ndc80, Cks2, Nuf2, Cks1b, Mki67, Tmpo, Cenpf,
Tacc3, Pimreg, Smc4, Ccnb2, Ckap2l, Ckap2, Aurkb, Bub1, Kif11, Anp32e, Tubb4b, Gtse1, Kif20b,
Hjurp, Cdca3, Jpt1, Cdc20, Ttk, Cdc25c, Kif2c, Rangap1, Ncapd2, Dlgap5, Cdca2, Cdca8, Ect2, Kif23,
Hmmr, Aurka, Psrc1, Anln, Lbr, Ckap5, Cenpe, Ctcf, Nek2, G2e3, Gas2l3, Cbx5, Cenpa}

\item{species}{The species of the data, either "human" or "mouse". Default is "auto".
If "auto" is given, the function will try to guess the species based on the gene names in the data.}

\item{log}{Logger}
}
\value{
The Seurat object with the cell cycle scores and predicted phase added to the metadata.
Literally, columns "S.Score", "G2M.Score" and "Phase" will be added to the metadata.
}
\description{
This function will run Seurat's CellCycleScoring on the given Seurat object and add the scores and predicted cell cycle phase to the metadata.
It also supports mouse data.
}
