% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-macros.R
\name{RunSeuratClustering}
\alias{RunSeuratClustering}
\title{Run seurat unsupervised clustering}
\usage{
RunSeuratClustering(
  object,
  use_sct = FALSE,
  SCTransformArgs = list(),
  NormalizeDataArgs = list(),
  FindVariableFeaturesArgs = list(),
  ScaleDataArgs = list(),
  RunPCAArgs = list(),
  UMAPArgs = list(),
  FindNeighborsArgs = list(),
  FindClustersArgs = list(),
  log = NULL,
  cache = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{use_sct}{Whether to use \code{\link[Seurat:SCTransform]{Seurat::SCTransform()}}}

\item{SCTransformArgs}{Arguments to pass to \code{\link[Seurat:SCTransform]{Seurat::SCTransform()}}}

\item{NormalizeDataArgs}{Arguments to pass to \code{\link[Seurat:NormalizeData]{Seurat::NormalizeData()}}}

\item{FindVariableFeaturesArgs}{Arguments to pass to \code{\link[Seurat:FindVariableFeatures]{Seurat::FindVariableFeatures()}}}

\item{ScaleDataArgs}{Arguments to pass to \code{\link[Seurat:ScaleData]{Seurat::ScaleData()}}}

\item{RunPCAArgs}{Arguments to pass to \code{\link[Seurat:RunPCA]{Seurat::RunPCA()}}}

\item{UMAPArgs}{Arguments to pass to \code{\link[Seurat:RunUMAP]{Seurat::RunUMAP()}}}

\item{FindNeighborsArgs}{Arguments to pass to \code{\link[Seurat:FindNeighbors]{Seurat::FindNeighbors()}}}

\item{FindClustersArgs}{Arguments to pass to \code{\link[Seurat:FindClusters]{Seurat::FindClusters()}}}

\item{log}{Logger}

\item{cache}{Directory to cache the results. Set to \code{FALSE} to disable caching}
}
\value{
The Seurat object with clustering results
}
\description{
Run seurat unsupervised clustering
}
