% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-macros.R
\name{RunSeuratDEAnalysis}
\alias{RunSeuratDEAnalysis}
\title{RunSeuratDEAnalysis}
\usage{
RunSeuratDEAnalysis(
  object,
  group_by,
  ident_1 = NULL,
  ident_2 = NULL,
  assay = NULL,
  subset = NULL,
  cache = NULL,
  error = TRUE,
  log = NULL,
  log_prefix = "",
  ...
)
}
\arguments{
\item{object}{Seurat object}

\item{group_by}{Column name in meta.data to group cells by}

\item{ident_1}{Identity of the first group of cells}

\item{ident_2}{Identity of the second group of cells}

\item{assay}{Assay to use for analysis}

\item{subset}{Subset of cells to use for analysis
It should be a string of expression to pass to \code{dplyr::filter} function}

\item{cache}{Directory to cache the results. Set to \code{FALSE} to disable caching
The results of the analysis will be saved in this directory
Anything changed in the object or arguments will trigger a re-run}

\item{error}{Whether to raise an error if the analysis fails
Otherwise, return an empty data frame}

\item{...}{Additional arguments to pass to \code{\link[Seurat:FindMarkers]{Seurat::FindMarkers()}}}
}
\description{
Run differential expression analysis on a Seurat object
}
\examples{
RunSeuratDEAnalysis(SeuratObject::pbmc_small, "groups", "g1", "g2")
}
