% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-macros.R
\name{RunSeuratIntegration}
\alias{RunSeuratIntegration}
\title{Run data integration on Seurat object}
\usage{
RunSeuratIntegration(
  object,
  no_integration = FALSE,
  IntegrateLayersArgs = list(),
  log = NULL,
  cache = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{no_integration}{Whether to skip integration, just join layers}

\item{IntegrateLayersArgs}{Arguments to pass to \link[Seurat:IntegrateLayers]{Seurat::IntegrateLayers}}

\item{log}{Logger}

\item{cache}{Directory to cache the results. Set to \code{FALSE} to disable caching}
}
\value{
The Seurat object with integrated data
}
\description{
Run data integration on Seurat object
}
