% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-macros.R
\name{RunSeuratMap2Ref}
\alias{RunSeuratMap2Ref}
\title{Run Seurat MapQuery to reference}
\usage{
RunSeuratMap2Ref(
  object,
  ref,
  use,
  ident = "seurat_clusters",
  refnorm = c("auto", "LogNormalize", "SCT", "SCTransform"),
  skip_if_normalized = TRUE,
  split_by = NULL,
  ncores = 1,
  MapQuery_args = list(refdata = list()),
  FindTransferAnchors_args = list(),
  SCTransform_args = list(),
  NormalizeData_args = list(),
  log = NULL,
  cache = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{ref}{Reference Seurat object or a file path to a Seurat object with .rds or .h5seurat extension}

\item{use}{The name in the metadata of the reference to use as identity for the query after mapping}

\item{ident}{The name of the identity in metadata after mapping}

\item{refnorm}{The normalization method used by the reference.
The same normalization method will be used for the query.
\itemize{
\item "auto": automatically detect the normalization method used by the reference
\item "LogNormalize": LogNormalize
\item "SCT": SCTransform
\item "SCTransform": SCTransform (alias for SCT)
}}

\item{skip_if_normalized}{Skip normalization if the query is already normalized with the same method as the reference}

\item{split_by}{The name of the metadata to split the query object by and do the mapping separately}

\item{ncores}{Number of cores to use for parallel processing for the split query objects}

\item{MapQuery_args}{Arguments to pass to \code{\link[Seurat:MapQuery]{Seurat::MapQuery()}}.
The \code{use} argument will be added to the \code{refdata} list.}

\item{FindTransferAnchors_args}{Arguments to pass to \code{\link[Seurat:FindTransferAnchors]{Seurat::FindTransferAnchors()}}.}

\item{SCTransform_args}{Arguments to pass to \code{\link[Seurat:SCTransform]{Seurat::SCTransform()}}.
Will be used to normalize the query object if \code{refnorm} is set to "SCT"}

\item{NormalizeData_args}{Arguments to pass to \code{\link[Seurat:NormalizeData]{Seurat::NormalizeData()}}.
Will be used to normalize the query object if \code{refnorm} is set to "LogNormalize"}

\item{log}{Logger}

\item{cache}{Directory to cache the results. Set to \code{FALSE} to disable caching}
}
\value{
The Seurat object with the mapped data
}
\description{
This will normalize the query object, find transfer anchors, and map the query to the reference.
}
\seealso{
https://satijalab.org/seurat/articles/integration_mapping.html
}
