% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-macros.R
\name{RunSeuratScDblFinder}
\alias{RunSeuratScDblFinder}
\title{Run SCDblFinder on a Seurat object}
\usage{
RunSeuratScDblFinder(object, ncores = 1, ...)
}
\arguments{
\item{object}{Seurat object}

\item{ncores}{Number of cores to use}

\item{...}{Additional arguments to pass to \link[scDblFinder:scDblFinder]{scDblFinder::scDblFinder}}
}
\value{
The Seurat object with doublet detection results in \verb{@misc$scDblFinder}
}
\description{
Run SCDblFinder on a Seurat object
}
\keyword{internal}
