% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-macros.R
\name{RunSeuratTransformation}
\alias{RunSeuratTransformation}
\title{Run transformations on a Seurat object}
\usage{
RunSeuratTransformation(
  object,
  use_sct = FALSE,
  SCTransformArgs = list(),
  NormalizeDataArgs = list(),
  FindVariableFeaturesArgs = list(),
  ScaleDataArgs = list(),
  RunPCAArgs = list(),
  log = NULL,
  from_ccs = FALSE,
  cache = NULL
)
}
\arguments{
\item{object}{Seurat object}

\item{use_sct}{Whether to use \link[Seurat:SCTransform]{Seurat::SCTransform}}

\item{SCTransformArgs}{Arguments to pass to \link[Seurat:SCTransform]{Seurat::SCTransform}}

\item{NormalizeDataArgs}{Arguments to pass to \link[Seurat:NormalizeData]{Seurat::NormalizeData}}

\item{FindVariableFeaturesArgs}{Arguments to pass to \link[Seurat:FindVariableFeatures]{Seurat::FindVariableFeatures}}

\item{ScaleDataArgs}{Arguments to pass to \link[Seurat:ScaleData]{Seurat::ScaleData}
You can use \code{features = "__all__"} to scale all features, which will be expanded to all features in the object.}

\item{RunPCAArgs}{Arguments to pass to \link[Seurat:RunPCA]{Seurat::RunPCA}}

\item{log}{Logger}

\item{from_ccs}{Whether this is called from \link{RunSeuratCellCycleScoring}.}

\item{cache}{Directory to cache the results. Set to \code{FALSE} to disable caching}
}
\value{
The transformed Seurat object
}
\description{
Run transformations on a Seurat object
}
\examples{
\donttest{
RunSeuratTransformation(SeuratObject::pbmc_small)
}
}
