% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-macros.R
\name{RunSeuratUMAP}
\alias{RunSeuratUMAP}
\title{Run seurat UMAP}
\usage{
RunSeuratUMAP(object, RunUMAPArgs = list(), cache = NULL, log = NULL)
}
\arguments{
\item{object}{Seurat object}

\item{RunUMAPArgs}{Arguments to pass to \code{\link[Seurat:RunUMAP]{Seurat::RunUMAP()}}.
\code{RunUMAPArgs$features} can be a character vector of features directly used for UMAP,
or a list with the following fields:
\itemize{
\item \code{order}: The order of the markers to use for UMAP, e.g. "desc(abs(avg_log2FC))"
\item \code{n}: The number of total features to use for UMAP, e.g. 30
If \code{RunUMAPArgs$features} is a list, it will run \code{\link[=RunSeuratDEAnalysis]{RunSeuratDEAnalysis()}} to get the markers
for each group, and then select the top \code{n}/\code{ngroups} features for each group
based on the \code{order} field.
If \code{RunUMAPArgs$features} is a numeric value, it will be treated as the \code{n} field
in the list above, with the default \code{order} being "desc(abs(avg_log2FC))".
\code{RunUMAPArgs$from} can be a path to a TSV file containing pre-computed UMAP
coordinates to overwrite the results of \code{\link[Seurat:RunUMAP]{Seurat::RunUMAP()}}. The path should be
in the format \code{"file:///path/to/file#id_col,coord_col1,coord_col2"}, where
\code{id_col} is the column with cell barcodes and the remaining columns are the UMAP
coordinate columns (e.g., \verb{UMAP_1,UMAP_2}). The cell barcodes must match those
in the Seurat object, and the coordinate column names must match the reduction's
\code{cell.embeddings} column names. You can also use 1-based column indices instead of names, e.g., \code{"file:///path/to/file#1,2,3"}.
}}

\item{cache}{Directory to cache the all markers, which can be reused later if
DE analysis is desired.}

\item{log}{The logger to use. If NULL, a default logger will be used.}
}
\value{
The Seurat object with UMAP results
}
\description{
In additional to \code{\link[Seurat:RunUMAP]{Seurat::RunUMAP()}}, we provide an additional arguments to use
markers as features for UMAP, which makes the UMAP more separated for different clusters.
}
\details{
When both \code{RunUMAPArgs$features} and \code{RunUMAPArgs$dims} are provided,
\code{RunUMAPArgs$dims} will be ignored.
If neither \code{RunUMAPArgs$features} nor \code{RunUMAPArgs$dims} is provided,
\code{RunUMAPArgs$dims} will be set to \code{1:min(30, ceiling(ncells/3), ncol(object@reductions[[reduction]]))},
where \code{ncells} is the number of cells in the object, and \code{reduction} is \code{RunUMAPArgs$reduction} (default: "pca").
}
