% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{slugify}
\alias{slugify}
\title{Slugify strings}
\usage{
slugify(
  x,
  non_alphanum_replace = "-",
  collapse_replace = TRUE,
  tolower = FALSE
)
}
\arguments{
\item{x}{strings to slugify}

\item{non_alphanum_replace}{Replace non-alphanumeric characters}

\item{collapse_replace}{Collapse consecutive non-alphanumeric character replacements}

\item{tolower}{Convert to lowercase}
}
\value{
slugified strings
}
\description{
Slugify strings
}
