% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrna-viz.R
\name{VizDEGs}
\alias{VizDEGs}
\title{Visualize DEGs}
\usage{
VizDEGs(
  degs,
  plot_type = c("volcano_pct", "volcano_log2fc", "violin", "box", "bar", "ridge", "dim",
    "heatmap", "dot"),
  order_by = "desc(abs(avg_log2FC))",
  genes = 10,
  outprefix = NULL,
  devpars = list(res = 100),
  more_formats = c(),
  save_code = FALSE,
  show_row_names = TRUE,
  show_column_names = TRUE,
  ...
)
}
\arguments{
\item{degs}{DEGs from RunSeuratDEAnalysis}

\item{plot_type}{Type of plot to generate
One of 'volcano_pct', 'volcano_log2fc', 'violin', 'box', 'bar', 'ridge', 'dim', 'heatmap', 'dot'}

\item{order_by}{An expression in string to order the genes}

\item{genes}{Number of genes genes to visualize (based on the 'order_by' expression)
Or an expression in string to filter the genes (passed by \link[dplyr:filter]{dplyr::filter})
Only works when plot_type is not a volcano plot}

\item{outprefix}{Prefix of the output file}

\item{devpars}{List of parameters to save the plot}

\item{more_formats}{Additional formats to save the plot in addition to 'png'}

\item{save_code}{Whether to save the code to reproduce the plot}

\item{show_row_names}{Whether to show row names in the heatmap}

\item{show_column_names}{Whether to show column names in the heatmap}

\item{...}{Additional arguments to pass to the plot function
\itemize{
\item For 'volcano_pct' and 'volcano_log2fc', additional arguments to pass to 'scplotter::VolcanoPlot'
\item For 'violin', 'box', 'bar', 'ridge', 'dim', 'heatmap', 'dot', additional arguments to pass to 'scplotter::FeatureStatPlot'
}}
}
\value{
A ggplot object if 'outprefix' is NULL, otherwise, save the plot to the output directory
}
\description{
Visualize differentially expressed genes
}
\examples{
\donttest{
degs <- suppressWarnings(RunSeuratDEAnalysis(SeuratObject::pbmc_small, "groups", "g1", "g2"))
VizDEGs(degs, plot_type = "volcano_pct")
VizDEGs(degs, plot_type = "volcano_log2fc")
VizDEGs(degs, plot_type = "violin")
VizDEGs(degs, plot_type = "box")
VizDEGs(degs, plot_type = "bar")
VizDEGs(degs, plot_type = "ridge")
VizDEGs(degs, plot_type = "dim")
# VizDEGs(degs, plot_type = "heatmap")
# VizDEGs(degs, plot_type = "dot")
}
}
